Differential operators

Docstrings

SphericalFunctions.Lz — Method

Lz(s, ℓₘᵢₙ, ℓₘₐₓ, [T])

Compute the angular-momentum operator associated with the $z$ direction. This is the standard $L_z$ operator, familiar from basic physics, extended to work with SWSHs. Note that this is the left Lie derivative; see Rz for the equivalent right Lie derivative. See the documentation or Boyle for more details.

In terms of the SWSHs, we can write the action of $L_z$ as

\[L_z {}_{s}Y_{ℓ,m} = m\, {}_{s}Y_{ℓ,m}\]

See also L², L₊, L₋, R², Rz, R₊, R₋, ð, ð̄.

SphericalFunctions.L² — Method

L²(s, ℓₘᵢₙ, ℓₘₐₓ, [T])

Compute the total angular-momentum operator from ℓₘᵢₙ up to ℓₘₐₓ. If not present, ℓₘᵢₙ is assumed to be abs(s). The argument s is ignored; it is present only for consistency with other operators, and is assumed to be 0 if not present.

This is the standard $L²$ operator, familiar from basic physics, extended to work with SWSHs. It is also known as the Casimir operator, and is equal to

\[L^2 = L_x^2 + L_y^2 + L_z^2 = \frac{L_+L_- + L_-L_+ + 2L_zL_z}{2}.\]

Note that these are the left Lie derivatives, but $L^2 = R^2$, where R is the right Lie derivative. See the documentation or Boyle for more details.

In terms of the SWSHs, we can write the action of $L^2$ as

\[L^2 {}_{s}Y_{ℓ,m} = ℓ\,(ℓ+1) {}_{s}Y_{ℓ,m}\]

See also Lz, L₊, L₋, R², Rz, R₊, R₋, ð, ð̄.

SphericalFunctions.L₊ — Method

L₊(s, ℓₘᵢₙ, ℓₘₐₓ, [T])

Compute the angular-momentum raising operator. This is the standard $L_+$ operator, familiar from basic physics, extended to work with SWSHs. Note that this is the left Lie derivative; see R₊ for the equivalent right Lie derivative. See the documentation or Boyle for more details.

We define $L_+$ to be the raising operator for the left Lie derivative with respect to rotation about $z$: $L_z$. By definition, this implies the commutator relation $[L_z, L_+] = L_+$, which allows us to derive $L_+ = L_x + i\, L_y.$

In terms of the SWSHs, we can write the action of $L_+$ as

\[L_+ {}_{s}Y_{ℓ,m} = \sqrt{(ℓ-m)(ℓ+m+1)}\, {}_{s}Y_{ℓ,m+1}.\]

Consequently, the mode weights of a function are affected as

\[\left\{L_+(f)\right\}_{s,ℓ,m} = \sqrt{(ℓ+m)(ℓ-m+1)}\,\left\{f\right\}_{s,ℓ,m-1}.\]

See also L², Lz, L₋, R², Rz, R₊, R₋, ð, ð̄.

SphericalFunctions.L₋ — Method

L₋(s, ℓₘᵢₙ, ℓₘₐₓ, [T])

Compute the angular-momentum lowering operator. This is the standard $L_-$ operator, familiar from basic physics, extended to work with SWSHs. Note that this is the left Lie derivative; see R₋ for the equivalent right Lie derivative. See the documentation or Boyle for more details.

We define $L_-$ to be the lowering operator for the left Lie derivative with respect to rotation about $z$: $L_z$. By definition, this implies the commutator relation $[L_z, L_-] = -L_-$, which allows us to derive $L_- = L_x - i\, L_y.$

In terms of the SWSHs, we can write the action of $L_-$ as

\[L_- {}_{s}Y_{ℓ,m} = \sqrt{(ℓ+m)(ℓ-m+1)}\, {}_{s}Y_{ℓ,m-1}.\]

Consequently, the mode weights of a function are affected as

\[\left\{L_-(f)\right\}_{s,ℓ,m} = \sqrt{(ℓ-m)(ℓ+m+1)}\,\left\{f\right\}_{s,ℓ,m+1}.\]

See also L², Lz, L₊, L₋, R², Rz, R₊, R₋, ð, ð̄.

SphericalFunctions.Rz — Method

Rz(s, ℓₘᵢₙ, ℓₘₐₓ, [T])

Compute the right angular-momentum operator associated with the $z$ direction.

This is the $R_z$ operator, much like the $L_z$ operator familiar from basic physics, but in terms of the right Lie derivative, and extended to work with SWSHs. See Lz for the equivalent left Lie derivative. See the documentation or Boyle for more details.

In terms of the SWSHs, we can write the action of $R_z$ as

\[R_z {}_{s}Y_{ℓ,m} = -s\, {}_{s}Y_{ℓ,m}\]

Note the unfortunate sign of $s$, which seems to be opposite to what we expect, and arises from the choice of definition of $s$ in the original paper by Newman and Penrose.

See also L², Lz, L₊, L₋, R², R₊, R₋, ð, ð̄.

SphericalFunctions.R² — Method

R²(s, ℓₘᵢₙ, ℓₘₐₓ, [T])

Compute the total angular-momentum operator from ℓₘᵢₙ up to ℓₘₐₓ. If not present, ℓₘᵢₙ is assumed to be abs(s). The argument s is ignored; it is present only for consistency with other operators, and is assumed to be 0 if not present.

This is the $R^2$ operator, much like the $L^2$ operator familiar from basic physics, but in terms of the right Lie derivative, and extended to work with SWSHs. It is also known as the Casimir operator, and is equal to

\[R^2 = R_x^2 + R_y^2 + R_z^2 = \frac{R_+R_- + R_-R_+ + 2R_zR_z}{2}.\]

Note that these are the right Lie derivatives, but $L^2 = R^2$, where L is the left Lie derivative. See the documentation or Boyle for more details.

In terms of the SWSHs, we can write the action of $R^2$ as

\[R^2 {}_{s}Y_{ℓ,m} = ℓ\,(ℓ+1) {}_{s}Y_{ℓ,m}\]

See also L², Lz, L₊, L₋, Rz, R₊, R₋, ð, ð̄.

SphericalFunctions.R₊ — Method

R₊(s, ℓₘᵢₙ, ℓₘₐₓ, [T])

Compute the right angular-momentum raising operator.

This is the $R_+$ operator, much like the $L_+$ operator familiar from basic physics, but in terms of the right Lie derivative, and extended to work with SWSHs. See L₊ for the equivalent left Lie derivative. See the documentation or Boyle for more details.

We define $R_+$ to be the raising operator for the right Lie derivative with respect to rotation about $z$: $R_z$. By definition, this implies the commutator relation $[R_z, R_+] = R_+$, which allows us to derive $R_+ = R_x - i\, R_y.$

In terms of the SWSHs, we can write the action of $R_+$ as

\[R_+ {}_{s}Y_{ℓ,m} = \sqrt{(ℓ+s)(ℓ-s+1)}\, {}_{s-1}Y_{ℓ,m}\]

Consequently, the mode weights of a function are affected as

\[\left\{R_+(f)\right\}_{s,ℓ,m} = \sqrt{(ℓ+s)(ℓ-s+1)}\,\left\{f\right\}_{s-1,ℓ,m}.\]

Because of the unfortunate sign of $s$ arising from the choice of definition of $s$ in the original paper by Newman and Penrose, this is a lowering operator for $s$, though it really is a raising operator for $R_z$, and raises the eigenvalue of the corresponding Wigner matrix.

See also L², Lz, L₊, L₋, R², Rz, R₋, ð, ð̄.

SphericalFunctions.R₋ — Method

R₋(s, ℓₘᵢₙ, ℓₘₐₓ, [T])

Compute the right angular-momentum lowering operator.

This is the $R_-$ operator, much like the $L_-$ operator familiar from basic physics, but in terms of the right Lie derivative, and extended to work with SWSHs. See L₋ for the equivalent left Lie derivative. See the documentation or Boyle for more details.

We define $R_-$ to be the raising operator for the right Lie derivative with respect to rotation about $z$: $R_z$. By definition, this implies the commutator relation $[R_z, R_-] = -R_-$, which allows us to derive $R_- = R_x + i\, R_y.$

In terms of the SWSHs, we can write the action of $R_-$ as

\[R_- {}_{s}Y_{ℓ,m} = \sqrt{(ℓ-s)(ℓ+s+1)}\, {}_{s+1}Y_{ℓ,m}\]

Consequently, the mode weights of a function are affected as

\[\left\{R_-(f)\right\}_{s,ℓ,m} = \sqrt{(ℓ-s)(ℓ+s+1)}\,\left\{f\right\}_{s+1,ℓ,m}.\]

Because of the unfortunate sign of $s$ arising from the choice of definition of $s$ in the original paper by Newman and Penrose, this is a raising operator for $s$, though it really is a lowering operator for $R_z$, and lowers the eigenvalue of the corresponding Wigner matrix - though that raises the eigenvalue of the corresponding Wigner matrix.

See also L², Lz, L₊, L₋, R², Rz, R₊, ð, ð̄.

SphericalFunctions.ð — Method

ð(s, ℓₘᵢₙ, ℓₘₐₓ, [T])

Compute coefficients for the spin-raising operator $\eth$.

This operator was originally defined by Newman and Penrose, but is more completely defined by Boyle. It is identical to R₋. Refer to that function's documentation for more details.

By definition, the spin-raising operator satisfies the commutator relation $[S, \eth] = \eth$ (where $S$ is the spin operator, which just multiplies the function by its spin). In terms of the SWSHs, we can write the action of $\eth$ as

\[ \eth {}_{s}Y_{ℓ,m} = \sqrt{(ℓ-s)(ℓ+s+1)} {}_{s+1}Y_{ℓ,m}.\]

Consequently, the mode weights of a function are affected as

\[\left\{\eth f\right\}_{s,ℓ,m} = \sqrt{(ℓ-s)(ℓ+s+1)}\,\left\{f\right\}_{s+1,ℓ,m}.\]

See also ð̄, L², Lz, L₊, L₋, R², Rz, R₊.

SphericalFunctions.ð̄ — Method

ð̄(s, ℓₘᵢₙ, ℓₘₐₓ, [T])

Compute coefficients for the spin-lowering operator $\bar{\eth}$.

This operator was originally defined by Newman and Penrose, but is more completely defined by Boyle. It is opposite to R₊ — meaning that $\bar{\eth} = -R₊$. Refer to that function's documentation for more details.

By definition, the spin-lowering operator satisfies the commutator relation $[S, \bar{\eth}] = -\bar{\eth}$ (where $S$ is the spin operator, which just multiplies the function by its spin). In terms of the SWSHs, we can write the action of $\bar{\eth}$ as

\[\bar{\eth} {}_{s}Y_{ℓ,m} = -\sqrt{(ℓ+s)(ℓ-s+1)} {}_{s-1}Y_{ℓ,m}.\]

Consequently, the mode weights of a function are affected as

\[\left\{\bar{\eth} f\right\}_{s,ℓ,m} = -\sqrt{(ℓ-s)(ℓ+s+1)}\,\left\{f\right\}_{s+1,ℓ,m}.\]

See also ð, L², Lz, L₊, L₋, R², Rz, R₊.