Wigner's $𝔇$ and $d$ matrices

Wigner's $𝔇$ matrices — and to a lesser extent, the related $d$ matrices — are extremely important in the theory of rotations. Each element is, itself, a special function of the rotation group: in particular, an eigenfunction of the left- and right-Lie derivatives, and thus a spin-weighted spherical function. Collectively, they describe how spin-weighted spherical functions transform under rotation. But their accurate and efficient computation is surprisingly subtle. This package implements the current state-of-the-art techniques for their fast and accurate computation, based on the $H$ recursion.

The actual computations can be done with the D_matrices function:

using Quaternionic
using SphericalFunctions
using SphericalFunctions.Deprecated

R = randn(RotorF64)
ℓₘₐₓ = 8
𝔇 = D_matrices(R, ℓₘₐₓ)

However, the matrices can take up a lot of memory. So for maximum efficiency when calling this function repeatedly with different R values, it is best to pre-allocate the necessary memory with the D_prep function, and the pass that in as an argument to D_matrices!:

D_storage = D_prep(ℓₘₐₓ)
𝔇 = D_matrices!(D_storage, R)

(Beware that, as noted in the documentation for D_matrices!, the output 𝔇 is just a reference to part of the D_storage object, so you should not reuse D_storage until you have copied or otherwise finished using 𝔇.)

The output 𝔇 is a (linear!) vector of Complex numbers with the same base type as R. The ordering of the elements is described in the documentation for D_matrices. It is also possible to efficiently view slices of this vector as a series of individual matrices using a D_iterator:

for (ℓ, Dˡ) in zip(0:ℓₘₐₓ, D_iterator(𝔇, ℓₘₐₓ))
    # Do something with the matrix Dˡ[ℓ+m′+1, ℓ+m+1]
end

For the $d$ matrices, we have almost the same interface, except that instead of the input quaternion R we only need the angle β (or its complex angle expiβ, which can be computed directly in some cases), and the output is real-valued:

using Quaternionic
using SphericalFunctions
using SphericalFunctions.Deprecated

β = π * rand(Float64)
ℓₘₐₓ = 8
d = d_matrices(β, ℓₘₐₓ)

Again, for repeated calls, it is best to pre-allocate storage:

d_storage = d_prep(ℓₘₐₓ)
d = d_matrices!(d_storage, β)

The output d is a vector of numbers of the same type as β, ordered in the same way as the output of D_matrices. And similarly, we can iterate over the individual matrices using a d_iterator.

Docstrings

SphericalFunctions.Deprecated.D_matrices — Function

D_matrices(R, ℓₘₐₓ)
D_matrices(α, β, γ, ℓₘₐₓ)

Compute Wigner's 𝔇 matrices $𝔇^{(ℓ)}_{m',m}(β)$ for all $ℓ \leq ℓ_\mathrm{max}$.

See D_matrices! for details about the input and output values.

This function only appropriate when you need to evaluate the matrices for a single value of R or α, β, γ because it allocates large arrays and performs many calculations that could be reused. If you need to evaluate the matrices for many values of R or α, β, γ, you should pre-allocate the storage with D_prep, and then call D_matrices! with the result instead.

SphericalFunctions.Deprecated.D_matrices! — Function

D_matrices!(D_storage, R)
D_matrices!(D_storage, α, β, γ)
D_matrices!(D, R, ℓₘₐₓ)
D_matrices!(D, α, β, γ, ℓₘₐₓ)

Compute Wigner's 𝔇 matrices $𝔇^{(ℓ)}_{m',m}(β)$ for all $ℓ \leq ℓ_\mathrm{max}$.

In all cases, the result is returned in a 1-dimensional array ordered as

[
    𝔇ˡₘₚ,ₘ(R)
    for ℓ ∈ 0:ℓₘₐₓ
    for mp ∈ -ℓ:ℓ
    for m ∈ -ℓ:ℓ
]

When the first argument is D, it will be modified, so it must be at least as large as that array. When the first argument is D_storage, it should be the quantity returned by D_prep, and the result will be written into the D field of that tuple. Both of these options — especially the latter — reduce the number of allocations needed on each call to the corresponding functions, which should increase the speed significantly. Note that the D or D_storage arguments must have types compatible with the type of R or α, β, γ.

Warn

When using the D_storage argument (which is recommended), the returned quantity D will be an alias of D_storage[1]. If you want to retain that data after the next call to D_matrices!, you should copy it with copy(D).

The α, β, γ arguments are Euler angles as described in the documentation of Quaternionic.from_euler_angles.

See also D_matrices for a simpler function call when you only need to evaluate the matrices for a single value of R or α, β, γ.

Examples

using Quaternionic, SphericalFunctions
ℓₘₐₓ = 8
T = Float64
R = Rotor{T}(1, 2, 3, 4)  # Will be normalized automatically
D_storage = D_prep(ℓₘₐₓ, T)
D = D_matrices!(D_storage, R)

SphericalFunctions.Deprecated.D_prep — Function

D_prep(ℓₘₐₓ, [T=Float64])

Construct storage space and pre-compute recursion coefficients to compute Wigner's $𝔇$ matrix in place.

This returns the D_storage arguments needed by D_matrices!.

SphericalFunctions.Deprecated.D_iterator — Type

D_iterator(D, ℓₘₐₓ, [ℓₘᵢₙ])

Construct an Iterator that returns sub-matrices of D, each of which consists of elements $(ℓ,-ℓ,-ℓ)$ through $(ℓ,ℓ,ℓ)$, for $ℓ$ from ℓₘᵢₙ through ℓₘₐₓ. By default, ℓₘᵢₙ is 0.

Inconsistent behavior

This iterator mistakenly returns the transpose of the result implied by this documentation. As a result, a warning is issued every time this function is called. Rather than actually fixing this bug in this minor/patch version — which would be a breaking change — this is a final release in this major version of the package to notify users of this function (and d_iterator) that there is a problem. The next major version of the package will likely change the actual behavior to the one implied by this docstring. To quiet these warnings, you can temporarily pass the keyword argument warn=false, though this will probably be removed in the next major version. Alternatively, use Dit = with_logger(NullLogger()) do D_iterator(...) end to catch any warnings.

Note that the returned objects are views into the original D data — meaning that you may alter their values.

Because the result is a matrix restricted to a particular $ℓ$ value, you can index the $(ℓ, m′, m)$ element as [ℓ+m′+1, ℓ+m+1]. For example, you might use this as something like

for (ℓ, Dˡ) in zip(ℓₘᵢₙ:ℓₘₐₓ, D_iterator(D, ℓₘₐₓ))
    for m′ in -ℓ:ℓ
        for m in -ℓ:ℓ
            Dˡ[ℓ+m′+1, ℓ+m+1]  # ... do something with Dˡ
        end
    end
end

Also note that no bounds checking is done, either at instantiation time or during iteration. You are responsible for ensuring that the size of D and the values of ℓₘₐₓ and ℓₘᵢₙ make sense.

SphericalFunctions.Deprecated.d_matrices — Function

d_matrices(β, ℓₘₐₓ)
d_matrices(expiβ, ℓₘₐₓ)

Compute Wigner's $d^{(ℓ)}$ matrices with elements $d^{(ℓ)}_{m',m}(β)$ for all $ℓ \leq ℓ_\mathrm{max}$. The $d$ matrices are sometimes called the "reduced" Wigner matrices, in contrast to the full $𝔇$ matrices.

See d_matrices! for details about the input and output values.

This function only appropriate when you need to evaluate the matrices for a single value of β or expiβ because it allocates large arrays and performs many calculations that could be reused. If you need to evaluate the matrices for many values of β or expiβ, you should pre-allocate the storage with d_prep, and then call d_matrices! with the result instead.

SphericalFunctions.Deprecated.d_matrices! — Function

d_matrices!(d_storage, β)
d_matrices!(d_storage, expiβ)
d_matrices!(d, β, ℓₘₐₓ)
d_matrices!(d, expiβ, ℓₘₐₓ)

Compute Wigner's $d^{(ℓ)}$ matrices with elements $d^{(ℓ)}_{m',m}(β)$ for all $ℓ \leq ℓ_\mathrm{max}$. The $d$ matrices are sometimes called the "reduced" Wigner matrices, in contrast to the full $𝔇$ matrices.

In all cases, the result is returned in a 1-dimensional array ordered as

[
    dˡₘₚ,ₘ(β)
    for ℓ ∈ 0:ℓₘₐₓ
    for mp ∈ -ℓ:ℓ
    for m ∈ -ℓ:ℓ
]

When the first argument is d, it will be modified, so it must be at least as large as that array. When the first argument is d_storage, it should be the quantity returned by d_prep, and the result will be written into the d field of that tuple. Both of these options — especially the latter — reduce the number of allocations needed on each call to the corresponding functions, which should increase the speed significantly.

Warn

When using the d_storage argument (which is recommended), the returned quantity d will be an alias of d_storage[1]. If you want to retain that data after the next call to d_matrices!, you should copy it with copy(d).

See also d_matrices for a simpler function call when you only need to evaluate the matrices for a single value of β or expiβ.

Examples

using SphericalFunctions
ℓₘₐₓ = 8
T = Float64
β = T(1)/8
d_storage = d_prep(ℓₘₐₓ, T)
d = d_matrices!(d_storage, β)

SphericalFunctions.Deprecated.d_prep — Function

d_prep(ℓₘₐₓ, [T=Float64])

Construct space and pre-compute recursion coefficients to compute Wigner's $d$ matrix in place.

This returns the d_storage arguments needed by d_matrices!.

SphericalFunctions.Deprecated.d_iterator — Type

d_iterator(d, ℓₘₐₓ, [ℓₘᵢₙ])

Construct an Iterator that returns sub-matrices of d, each of which consists of elements $(ℓ,-ℓ,-ℓ)$ through $(ℓ,ℓ,ℓ)$, for $ℓ$ from ℓₘᵢₙ through ℓₘₐₓ. By default, ℓₘᵢₙ is 0.

Inconsistent behavior

This iterator mistakenly returns the transpose of the result implied by this documentation. As a result, a warning is issued every time this function is called. Rather than actually fixing this bug in this minor/patch version — which would be a breaking change — this is a final release in this major version of the package to notify users of this function (and D_iterator) that there is a problem. The next major version of the package will likely change the actual behavior to the one implied by this docstring. To quiet these warnings, you can temporarily pass the keyword argument warn=false, though this will probably be removed in the next major version. Alternatively, use Dit = with_logger(NullLogger()) do D_iterator(...) end to catch any warnings.

Note that the returned objects are views into the original d data — meaning that you may alter their values.

Because the result is a matrix restricted to a particular $ℓ$ value, you can index the $(ℓ, m′, m)$ element as [ℓ+m′+1, ℓ+m+1]. For example, you might use this as something like

for (ℓ, dˡ) in zip(ℓₘᵢₙ:ℓₘₐₓ, d_iterator(d, ℓₘₐₓ))
    for m′ in -ℓ:ℓ
        for m in -ℓ:ℓ
            dˡ[ℓ+m′+1, ℓ+m+1]  # ... do something with dˡ
        end
    end
end

Also note that no bounds checking is done, either at instantiation time or during iteration. You are responsible for ensuring that the size of d and the values of ℓₘₐₓ and ℓₘᵢₙ make sense.

Workspaces

SphericalFunctions.HWedge — Type

HWedge{IT, RT, ST} <: AbstractWignerMatrix{IT, RT, ST}

The $Hˡ$ matrix is critical to efficient and stable computation of the Wigner $D$ and $d$ matrices — in fact, it essentially is the $d$ matrix with signs adjusted to avoid numerical problems with alternating signs. This gives it additional symmetries that reduce the amount of data that needs to be stored to about 1/4 of the total $d$ size.

The purpose of an HWedge is to provide a workspace for the Wigner recurrences that is efficient, both in terms of the size of memory used, and the implications for vectorization and threading. Specifically, the data is stored as strictly Real values, in contiguous storage. Indexing is performed efficiently via precomputed row offsets. Once the full recurrence is done, the data can be used directly — computing phases and symmetry on the fly — or copied into a full explicit matrix with the appropriate phases.

The recurrences require $m$ in the full range from 0 (or 1/2) to $ℓ$, but $m'$ only needs to include the axis $m'=0$ or 1/2. Thus, we store m′ₘᵢₙand m′ₘₐₓ as fields, and only require enough storage for those ranges. Specifically, an HWedge will store elements in a vector as if they were components of the Hˡ matrix:

[
    Hˡ[m′, m]
    for m′ ∈ max(-ℓ, m′ₘᵢₙ):min(ℓ, m′ₘₐₓ)
    for m ∈ abs(m′):ℓ
]

However, for further efficiency when vectorizing and threading over multiple rotors, the data is stored as a 1-dimensional vector, though it can be indexed as if it were a three-dimensional array, with the first dimension indexing Nᵣ different rotors, and the second dimension indexing m′, and the third dimension indexing m. Thus, this object can be indexed as Hˡ[iᵣ, m′, m] to get the Hˡ value for rotor index iᵣ, and matrix element (m′, m).

Because of this complicated layout, the constructor is fairly restrictive, but will do all the allocation needed. To avoid multiple allocations, it is advisable to first construct an instance with the maximum ℓ value that will be needed, and then change the ℓ field as needed to compute different orders. That is, if H isa HWedge, then H.ℓ = new_ell can be used to change the current order being computed. The constructor starts out with the smallest ℓ value possible (0 or 1/2), which is the natural choice for recurrence.

Thread safety

The HWedge object is not thread safe. Its internal storage is intended to be changed by different threads, but the code must be designed carefully to avoid accessing the same memory locations from different threads. In particular, note that changing the ℓ field changes internal storage, and is not thread-safe. It may be better to allocate separate HWedge objects for each thread.

SphericalFunctions.HAxis — Type

HAxis{IT, RT} <: AbstractWignerMatrix{IT, RT, FixedSizeVectorDefault{RT}}

The HAxis type represents the $m'=0$, $m≥0$ axis of the Hˡ matrix used in calculation of the Wigner D and d matrices.

As with HWedge, the data is stored as a 1-dimensional vector, though it can be indexed as if it were a two-dimensional array, with the first dimension indexing Nᵣ different rotors, and the second dimension indexing m — or alternatively as if it were a three-dimensional array with the second dimension indexing m′ and the third indexing m. Thus, this object can be indexed as Hˡ₀[iᵣ, m] or Hˡ[iᵣ, m′, m] to get the Hˡ value for rotor index iᵣ, and matrix element (m′, m).