Calculators

Typically, when calculating special functions, we will use recurrence relations along with some coefficients — which frequently requires significant setup processing. That processing can be cached, so that the calculations themselves consist primarily of memory accesses and simple arithmetic. For this reason, we use "calculator" objects, which will be constructed with some indication of the largest indices you expect to use. The calculators can then be called repeatedly for specific values of the arguments, which will compute the function values for all (or some subset of) indices.

Modules = [SphericalFunctions.AssociatedLegendreFunction]
Pages   = ["src/associated_legendre/calculator.jl"]

SphericalFunctions.D! — Method

D!(𝔇, R, ℓₘₐₓ, (aₙᵐ,bₙᵐ,dₙᵐ), expimα, expimγ)

Compute Wigner's 𝔇 matrix

This function implements the preferred method of computing the 𝔇 matrix for large ell values. In particular, above ell≈32 standard formulas become completely unusable because of numerical instabilities and overflow. This function uses stable recursion methods instead, and should be usable beyond ell≈1000.

This function computes 𝔇ˡₘₚ,ₘ(R). The result is returned in a 1-dimensional array ordered as

[
    𝔇(ell, mp, m, R)
    for ell in range(ell_max+1)
    for mp in range(-min(ℓ, mp_max), min(ℓ, mp_max)+1)
    for m in range(-ell, ell+1)
]

SphericalFunctions.Y! — Function

Y(s, R, ℓₘₐₓ)
Y!(Y, s, R, ℓₘₐₓ)

Evaluate (and write into Y, if present) the values of ${}_{s}Y_{\ell, m}(R)$ for the input value of s, for all $(\ell, m)$ throughout the range specified by wigner. R is assumed to be a unit quaternion (which may be Rotor, or simply a Quaternion). If R does not have unit magnitude, the output elements will be too large by a factor $|R|^{\ell}$. If Y is not present, a new array will be created.

The spherical harmonics of spin weight $s$ are related to Wigner's $\mathfrak{D}$ matrix as

\[\begin{aligned} {}_{s}Y_{\ell, m}(R) &= (-1)^s \sqrt{\frac{2\ell+1}{4\pi}} \mathfrak{D}^{(\ell)}_{m, -s}(R) \\ &= (-1)^s \sqrt{\frac{2\ell+1}{4\pi}} \bar{\mathfrak{D}}^{(\ell)}_{-s, m}(\bar{R}). \end{aligned}\]

SphericalFunctions.Yprep — Method

Yprep(ℓₘₐₓ, sₘₐₓ, T, ℓₘᵢₙ)

Prepare the storage, recursion coefficients, and workspace to compute ₛYₗ,ₘ data up to the maximum sizes given.

Returns a tuple of Y, H_rec_coeffs, Hwedge, expimϕ, which can be passed to the correspondingly named arguments of Y!.

Note that the same results of this function can be passed to Y!, even if the value of ℓₘₐₓ passed to that function is smaller than the value passed to this function, or the value of spin passed to that function is smaller (in absolute value) than the sₘₐₓ passed to this function. However, the value of ℓₘᵢₙ passed to that function must not be smaller than the value passed to this function (unless one of the other sizes is sufficiently smaller).

SphericalFunctions.d! — Method

d!(d, expiβ, ℓₘₐₓ, H_rec_coeffs)
d!(d, expiβ, ℓₘₐₓ)
d!(d, β, ℓₘₐₓ, H_rec_coeffs)
d!(d, β, ℓₘₐₓ)

Compute Wigner's d matrix dˡₘₚ,ₘ(β)

Notes

This function is the preferred method of computing the d matrix for large ell values. In particular, above ell≈32 standard formulas become completely unusable because of numerical instabilities and overflow. This function uses stable recursion methods instead, and should be usable beyond ell≈1000.

The result is returned in a 1-dimensional array ordered as

[
    d(ell, mp, m, β)
    for ell in range(ell_max+1)
    for mp in range(-min(ℓ, mp_max), min(ℓ, mp_max)+1)
    for m in range(-ell, ell+1)
]

Transformation

SphericalFunctions.map2salm! — Method

map2salm!(salm, map, spin, ℓmax)
map2salm!(salm, map, plan)

Transform map values sampled on the sphere to ${}_sa_{\ell, m}$ modes in place.

For details, see map2salm.

SphericalFunctions.map2salm — Method

map2salm(map, spin, ℓmax)
map2salm(map, plan)

Transform map values sampled on the sphere to ${}_sa_{\ell, m}$ modes.

The map array should have size Nφ along its first dimension and Nϑ along its second; any number of dimensions may follow. The spin must be entered explicitly, and ℓmax is the highest ℓ value you want in the output.

For repeated applications of this function with different values of map, it is more efficient to pre-compute plan using plan_map2salm. These functions will create a new salm array on each call. To operate in place on a pre-allocated salm array, use map2salm!.

The core of this function follows the method described by Reinecke and Seljebotn.

SphericalFunctions.plan_map2salm — Method

plan_map2salm(map, spin, ℓmax)

Precompute values to use in executing map2salm or map2salm!.

The arguments to this function exactly mirror those of the first form of map2salm, and all but the first argument in the first form of map2salm!. The plan returned by this function can be passed to the second forms of those functions to avoid some computation and allocation costs.

Note that the plan object is not thread safe; a separate plan should be created for each thread that will use one, or locks should be used to ensure that a single plan is not used at the same time on different threads.