Calculators

Typically, when calculating special functions, we will use recurrence relations along with some coefficients — which frequently requires significant setup processing. That processing can be cached, so that the calculations themselves consist primarily of memory accesses and simple arithmetic. For this reason, we use "calculator" objects, which will be constructed with some indication of the largest indices you expect to use. The calculators can then be called repeatedly for specific values of the arguments, which will compute the function values for all (or some subset of) indices.

SphericalFunctions.AssociatedLegendreFunction.ALFcompute! — Method

ALFcompute(expiβ, nmax)
ALFcompute!(P̄, expiβ, nmax)
ALFcompute(expiβ, nmax, recursion_coefficients)
ALFcompute!(P̄, expiβ, nmax, recursion_coefficients)

Compute the "fully normalized" Associated Legendre Functions up to some maximum n value nmax.

These functions can take a vector P̄, to store the data, stored in order of increasing m most rapidly varying and then increasing n. If not supplied, P̄ will be constructed for you and returned.

The optional recursion_coefficients argument must be an ALFRecursionCoefficients, which stores various constant coefficients used in the recursion. This object requires more than 3x the memory and more than 20x the time to compute a single P̄ vector without this argument, but passing it will typically speed up the calculation of each P̄ by a factor of 8x or so. Thus, if you expect to compute P̄ more than a few times, it will take less time to pre-compute those factors, and pass them to this function.

Note that the base real type will be inferred from the (complex) type of expiβ. If present, the base types of P̄ and recursion_coefficients must agree.

SphericalFunctions.WignerMatrixCalculator — Type

WignerMatrixCalculator(ℓmin, ℓmax, [m′max], [T])

Construct a WignerMatrixCalculator to use in calculating $𝔇$, $d$, or ${}_{s}Y_{ℓ,m}$.

If m′max is given, values may be calculated with ℓ ∈ ℓmin:ℓmax and m ∈ -ℓ:ℓ but only m′ ∈ ℓmin:m′max. This will reduce the storage and computation costs. Note that if your goal is just to compute spin-weighted spherical harmonics up to some spins, you should usem′max=abs(s)`.

If T<:Real is given, all calculations will be done using that type; the default is Float64.

See also:

D! to compute the full $𝔇$ matrix

d! to compute the small $d$ matrix

Y! to compute spin-weighted spherical harmonic values

SphericalFunctions.D! — Method

D!(𝔇, w, R)
D!(w, R)

Compute Wigner's 𝔇 matrix

Parameters

𝔇 : array_like, optional Array into which the 𝔇 values should be written. It should be an array of complex, with size self.Dsize. If not present, the array will be created. In either case, the array will also be returned.

workspace : optional A working array like the one returned by Wigner.new_workspace(). If not present, this object's default workspace will be used. Note that it is not safe to use the same workspace on multiple threads.

R : Quaternion Array to be interpreted as a quaternionic array (thus its final dimension must have size 4), representing the rotations on which the 𝔇 matrix will be evaluated.

Returns

D : array This is a 1-dimensional array of complex; see below.

See Also

H : Compute a portion of the H matrix

d : Compute the full Wigner d matrix

rotate : Avoid computing the full 𝔇 matrix and rotate modes directly

evaluate : Avoid computing the full 𝔇 matrix and evaluate modes directly

Notes

This function is the preferred method of computing the 𝔇 matrix for large ell values. In particular, above ell≈32 standard formulas become completely unusable because of numerical instabilities and overflow. This function uses stable recursion methods instead, and should be usable beyond ell≈1000.

This function computes 𝔇ˡₘₚ,ₘ(R). The result is returned in a 1-dimensional array ordered as

[
    𝔇(ell, mp, m, R)
    for ell in range(ell_max+1)
    for mp in range(-min(ℓ, mp_max), min(ℓ, mp_max)+1)
    for m in range(-ell, ell+1)
]

SphericalFunctions.Y! — Method

Y!(Y, wigner, s, R)
Y!(wigner, s, R)

Evaluate (and write into Y, if present) the values of ${}_{s}Y_{\ell, m}(R)$ for the input value of s, for all $(\ell, m)$ throughout the range specified by wigner. R is assumed to be a unit quaternion (which may be Rotor, or simply a Quaternion). If R does not have unit magnitude, the output elements will be too large by a factor $|R|^{\ell}$. If Y is not present, a new array will be created.

The spherical harmonics of spin weight $s$ are related to Wigner's $\mathfrak{D}$ matrix as

\[\begin{aligned} {}_{s}Y_{\ell, m}(R) &= (-1)^s \sqrt{\frac{2\ell+1}{4\pi}} \mathfrak{D}^{(\ell)}_{m, -s}(R) \\ &= (-1)^s \sqrt{\frac{2\ell+1}{4\pi}} \bar{\mathfrak{D}}^{(\ell)}_{-s, m}(\bar{R}). \end{aligned}\]

SphericalFunctions.d! — Method

d!(d, wigner, expiβ)

Compute Wigner's d matrix dˡₘₚ,ₘ(β)

Parameters

expiβ : array_like Values of expi(i*β) on which to evaluate the d matrix.

out : array_like, optional Array into which the d values should be written. It should be an array of floats, with size self.dsize. If not present, the array will be created. In either case, the array will also be returned.

workspace : arraylike, optional A working array like the one returned by Wigner.newworkspace(). If not present, this object's default workspace will be used. Note that it is not safe to use the same workspace on multiple threads.

Returns

d : array This is a 1-dimensional array of floats; see below.

See Also

H : Compute a portion of the H matrix

D : Compute the full Wigner 𝔇 matrix

rotate : Avoid computing the full 𝔇 matrix and rotate modes directly

evaluate : Avoid computing the full 𝔇 matrix and evaluate modes directly

Notes

This function is the preferred method of computing the d matrix for large ell values. In particular, above ell≈32 standard formulas become completely unusable because of numerical instabilities and overflow. This function uses stable recursion methods instead, and should be usable beyond ell≈1000.

The result is returned in a 1-dimensional array ordered as

[
    d(ell, mp, m, β)
    for ell in range(ell_max+1)
    for mp in range(-min(ℓ, mp_max), min(ℓ, mp_max)+1)
    for m in range(-ell, ell+1)
]

SphericalFunctions.sign — Method

Return sign of input, with sign(0)=1

Transformation

SphericalFunctions.map2salm! — Method

map2salm!(salm, map, spin, ℓmax, [ℓmin])
map2salm!(salm, map, plan)

Transform map values sampled on the sphere to ${}_sa_{\ell, m}$ modes in place.

For details, see map2salm.

SphericalFunctions.map2salm — Method

map2salm(map, spin, ℓmax, [ℓmin])
map2salm(map, plan)

Transform map values sampled on the sphere to ${}_sa_{\ell, m}$ modes.

The map array should have size Nφ along its first dimension and Nϑ along its second; any number of dimensions may follow. The spin must be entered explicitly, and ℓmax is the highest ℓ value you want in the output. The ℓmin represents the smallest ℓ value in the output and defaults to abs(spin); it generally should never be larger than this.

For repeated applications of this function with different values of map, it is more efficient to pre-compute plan using plan_map2salm. These functions will create a new salm array on each call. To operate in place on a pre-allocated salm array, use map2salm!.

The core of this function follows the method described by Reinecke and Seljebotn.

SphericalFunctions.plan_map2salm — Method

plan_map2salm(map, spin, ℓmax, [ℓmin])

Precompute values to use in executing map2salm or map2salm!.

The arguments to this function exactly mirror those of the first form of map2salm, and all but the first argument in the first form of map2salm!. The plan returned by this function can be passed to the second forms of those functions to avoid some computation and allocation costs.

Note that the plan object is not thread safe; a separate plan should be created for each thread that will use one, or locks should be used to ensure that a single plan is not used at the same time on different threads.