Calculators

Typically, when calculating special functions, we will use recurrence relations along with some coefficients — which frequently requires significant setup processing. That processing can be cached, so that the calculations themselves consist primarily of memory accesses and simple arithmetic. For this reason, we use "calculator" objects, which will be constructed with some indication of the largest indices you expect to use. The calculators can then be called repeatedly for specific values of the arguments, which will compute the function values for all (or some subset of) indices.

Spherical.AssociatedLegendreFunction.ALFcompute! — Method

ALFcompute(expiβ, nmax)
ALFcompute!(P̄, expiβ, nmax)
ALFcompute(expiβ, nmax, recursion_coefficients)
ALFcompute!(P̄, expiβ, nmax, recursion_coefficients)

Compute the "fully normalized" Associated Legendre Functions up to some maximum n value nmax.

These functions can take a vector P̄, to store the data, stored in order of increasing m most rapidly varying and then increasing n. If not supplied, P̄ will be constructed for you and returned.

The optional recursion_coefficients argument must be an ALFRecursionCoefficients, which stores various constant coefficients used in the recursion. This object requires more than 3x the memory and more than 20x the time to compute a single P̄ vector without this argument, but passing it will typically speed up the calculation of each P̄ by a factor of 8x or so. Thus, if you expect to compute P̄ more than a few times, it will take less time to pre-compute those factors, and pass them to this function.

Note that the base real type will be inferred from the (complex) type of expiβ. If present, the base types of P̄ and recursion_coefficients must agree.

Modules = [WignerMatrices]
Pages   = ["wigner_matrices/calculator.jl", "wigner_matrices/evaluate.jl"]

Transformation

Spherical.map2salm! — Method

map2salm!(salm, map, spin, ℓmax, [ℓmin])
map2salm!(salm, map, plan)

Transform map values sampled on the sphere to ${}_sa_{\ell, m}$ modes in place.

For details, see map2salm.

Spherical.map2salm — Method

map2salm(map, spin, ℓmax, [ℓmin])
map2salm(map, plan)

Transform map values sampled on the sphere to ${}_sa_{\ell, m}$ modes.

The map array should have size Nφ along its first dimension and Nϑ along its second; any number of dimensions may follow. The spin must be entered explicitly, and ℓmax is the highest ℓ value you want in the output. The ℓmin represents the smallest ℓ value in the output and defaults to abs(spin); it generally should never be larger than this.

For repeated applications of this function with different values of map, it is more efficient to pre-compute plan using plan_map2salm. These functions will create a new salm array on each call. To operate in place on a pre-allocated salm array, use map2salm!.

Spherical.plan_map2salm — Method

plan_map2salm(map, spin, ℓmax, [ℓmin])

Precompute values to use in executing map2salm or map2salm!.

The arguments to this function exactly mirror those of the first form map2salm.