Function calculators

Typically, when calculating special functions, we will use recursion relations along with some coefficients — which frequently requires significant setup processing. That processing can be cached, so that the calculations themselves consist primarily of memory accesses and simple arithmetic.

Similarly, some computations require a certain amount of "workspace" — significant amounts of memory that are needed for the computation, but are not part of the core result (and don't even need to be returned). Rather than allocating this memory on each call to the computation, we can allocate once and pass the workspace in as an argument.

These two patterns — pre-computed coefficients and pre-allocated workspaces — will appear as possible call signatures in each of the following functions. Note that we also have helper functions like Dprep, Dstorage, etc., which construct the necessary coefficients and/or workspaces, as well as the actual results for in-place computations, to make it easier to call the corresponding calculator functions.

Note that these functions impose assumptions about the ordering of arrays as vectors. For example, the series of Wigner $𝔇$ matrices will be stored as a single vector. And for particular values of $(ℓ, n, m)$, the element $𝔇ˡₙ,ₘ$ will have to be accessed as some linearly indexed element of that vector. These formats, along with utility functions for accessing particular elements, are described here.

The fundamental algorithm is the $H$ recursion, which is the core computation needed for Wigner's $d$ and $𝔇$ matrices, and the spin-weighted spherical harmonics ${}_{s}Y_{\ell,m}$, as well as map2salm functions.

SphericalFunctions.H! — Method

H!(H, expiβ, ℓₘₐₓ, m′ₘₐₓ, H_rec_coeffs)
H!(H, expiβ, ℓₘₐₓ, m′ₘₐₓ, H_rec_coeffs, Hindex)

Compute the $H$ matrix defined by Gumerov and Duraiswami.

This computation forms the basis for computing Wigner's $d$ and $𝔇$ matrices via d! and D!, the spin-weighted spherical harmonics via Y!, and for transforming from values of spin-weighted spherical functions evaluated on a grid to the corresponding mode weights via map2salm.

Due to symmetries, we only need to compute ~1/4 of the elements of this matrix, so only those elements with $m ≥ |m′|$ are computed. The relevant indices of the H vector are computed based on the Hindex function — which defaults to WignerHindex, but could reasonably be WignerDindex if the input H vector contains all valid indices. However, it is assumed that the storage scheme used for H is such that the successive $m$ values are located in successive elements.

If $m′ₘₐₓ < ℓₘₐₓ$, we don't even need 1/4 of the elements, and only values with $|m′| ≤ m′ₘₐₓ$ will be computed. This is particularly useful for computing spin-weighted spherical harmonics.

Note that the recursion coefficients H_rec_coeffs should be the quantity returned by H_recursion_coefficients.

SphericalFunctions.H_recursion_coefficients — Method

H_recursion_coefficients(ℓₘₐₓ, T)

Pre-compute constants used in Wigner H recursion.

Internally, the $H$ recursion relies on calculation of the Associated Legendre Functions (ALFs), which can also be called on their own:

SphericalFunctions.ALFcompute! — Method

ALFcompute(expiβ, nmax)
ALFcompute!(P̄, expiβ, nmax)
ALFcompute(expiβ, nmax, recursion_coefficients)
ALFcompute!(P̄, expiβ, nmax, recursion_coefficients)

Compute the "fully normalized" Associated Legendre Functions up to some maximum n value nmax.

These functions can take a vector P̄, to store the data, stored in order of increasing m most rapidly varying and then increasing n. If not supplied, P̄ will be constructed for you and returned.

The optional recursion_coefficients argument must be an ALFRecursionCoefficients, which stores various constant coefficients used in the recursion. This object requires more than 3x the memory and more than 20x the time to compute a single P̄ vector without this argument, but passing it will typically speed up the calculation of each P̄ by a factor of 8x or so. Thus, if you expect to compute P̄ more than a few times, it will take less time to pre-compute those factors, and pass them to this function.

Note that the base real type will be inferred from the (complex) type of expiβ. If present, the base types of P̄ and recursion_coefficients must agree.

Based on those, we have the d!, D!, and Y! functions:

SphericalFunctions.D! — Method

D!(𝔇, R, ℓₘₐₓ, (aₙᵐ,bₙᵐ,dₙᵐ), expimα, expimγ)

Compute Wigner's 𝔇 matrix

This function implements the preferred method of computing the 𝔇 matrix for large ell values. In particular, above ell≈32 standard formulas become completely unusable because of numerical instabilities and overflow. This function uses stable recursion methods instead, and should be usable beyond ell≈1000.

This function computes 𝔇ˡₘₚ,ₘ(R). The result is returned in a 1-dimensional array ordered as

[
    𝔇(ell, mp, m, R)
    for ell in range(ell_max+1)
    for mp in range(-min(ℓ, mp_max), min(ℓ, mp_max)+1)
    for m in range(-ell, ell+1)
]

SphericalFunctions.Dprep — Method

Dprep(ℓₘₐₓ, T)

Construct space and pre-compute recursion coefficients to compute Wigner's $𝔇$ matrix in place.

This returns the (D, H_rec_coeffs, expimα, expimγ) arguments needed by D!.

SphericalFunctions.Dstorage — Method

Dstorage(ℓₘₐₓ, T)

Construct space to compute Wigner's $𝔇$ matrix in place.

This returns the D argument needed by D!.

SphericalFunctions.Y! — Function

Y!(Y, R, ℓₘₐₓ, spin, H_rec_coeffs, Hwedge, expimϕ, ℓₘᵢₙ=0)
Y!(Y, expiθ, expiϕ, ℓₘₐₓ, spin, H_rec_coeffs, Hwedge, expimϕ, ℓₘᵢₙ=0)
Y!(Y, expiϕ, expiθ, expiγ, ℓₘₐₓ, spin, H_rec_coeffs, Hwedge, expimϕ, ℓₘᵢₙ=0)

Evaluate (and write into Y, if present) the values of ${}_{s}Y_{\ell, m}(R)$ for the input value of s, for all $(\ell, m)$ throughout the range specified by wigner. R is assumed to be a unit quaternion (which may be Rotor, or simply a Quaternion). If R does not have unit magnitude, the output elements will be too large by a factor $|R|^{\ell}$. If Y is not present, a new array will be created.

The spherical harmonics of spin weight $s$ are related to Wigner's $\mathfrak{D}$ matrix as

\[\begin{aligned} {}_{s}Y_{\ell, m}(R) &= (-1)^s \sqrt{\frac{2\ell+1}{4\pi}} \mathfrak{D}^{(\ell)}_{m, -s}(R) \\ &= (-1)^s \sqrt{\frac{2\ell+1}{4\pi}} \bar{\mathfrak{D}}^{(\ell)}_{-s, m}(\bar{R}). \end{aligned}\]

SphericalFunctions.Yprep — Method

Yprep(ℓₘₐₓ, sₘₐₓ, T, ℓₘᵢₙ)

Prepare the storage, recursion coefficients, and workspace to compute ₛYₗ,ₘ data up to the maximum sizes given.

Returns a tuple of Y, H_rec_coeffs, Hwedge, expimϕ, which can be passed to the correspondingly named arguments of Y!.

Note that the same results of this function can be passed to Y!, even if the value of ℓₘₐₓ passed to that function is smaller than the value passed to this function, or the value of spin passed to that function is smaller (in absolute value) than the sₘₐₓ passed to this function. However, the value of ℓₘᵢₙ passed to that function must not be smaller than the value passed to this function (unless one of the other sizes is sufficiently smaller).

SphericalFunctions.Ystorage — Method

Ystorage(ℓₘₐₓ, T)
Ystorage(ℓₘₐₓ, T, ℓₘᵢₙ)

Construct space to compute the spin-weighted spherical harmonics $ₛYₗ,ₘ$ in place.

This returns the Y argument needed by Y!.

SphericalFunctions.d! — Method

d!(d, expiβ, ℓₘₐₓ, H_rec_coeffs)
d!(d, expiβ, ℓₘₐₓ)
d!(d, β, ℓₘₐₓ, H_rec_coeffs)
d!(d, β, ℓₘₐₓ)
d(expiβ, ℓₘₐₓ)
d(β, ℓₘₐₓ)

Compute Wigner's d matrix dˡₘₚ,ₘ(β)

Notes

This function is the preferred method of computing the d matrix for large ell values. In particular, above ell≈32 standard formulas become completely unusable because of numerical instabilities and overflow. This function uses stable recursion methods instead, and should be usable beyond ell≈1000.

The result is returned in a 1-dimensional array ordered as

[
    d(ell, mp, m, β)
    for ell in range(ell_max+1)
    for mp in range(-min(ℓ, mp_max), min(ℓ, mp_max)+1)
    for m in range(-ell, ell+1)
]

SphericalFunctions.dprep — Method

dprep(ℓₘₐₓ, T)

Construct space and pre-compute recursion coefficients to compute Wigner's $d$ matrix in place.

This returns the (d, H_rec_coeffs) arguments needed by d!.

SphericalFunctions.dstorage — Method

dstorage(ℓₘₐₓ, T)

Construct space to compute Wigner's $d$ matrix in place.

This returns the d argument needed by d!.

In all cases, the returned (or overwritten) data are stored linearly, as a single Vector ranging over all indices. To access subarrays based on certain $ℓ$ values, for example, we also have iterators that return conveniently shaped views into these linear vectors:

SphericalFunctions.Diterator — Type

Diterator(D, ℓₘₐₓ, [ℓₘᵢₙ])

Construct an Iterator that returns sub-matrices of D, each of which consists of elements $(ℓ,-ℓ,-ℓ)$ through $(ℓ,ℓ,ℓ)$, for $ℓ$ from ℓₘᵢₙ through ℓₘₐₓ. By default, ℓₘᵢₙ is 0.

Note that the returned objects are views into the original D data — meaning that you may alter their values.

Because the result is a matrix restricted to a particular $ℓ$ value, you can index the $(ℓ, m′, m)$ element as [ℓ+m′+1, ℓ+m+1]. For example, you might use this as something like

for (ℓ, Dˡ) in zip(ℓₘᵢₙ:ℓₘₐₓ, Diterator(D, ℓₘₐₓ))
    for m′ in -ℓ:ℓ
        for m in -ℓ:ℓ
            Dˡ[ℓ+m′+1, ℓ+m+1] = <...>
        end
    end
end

Also note that no bounds checking is done, either at instantiation time or during iteration. You are responsible for ensuring that the size of D and the values of ℓₘₐₓ and ℓₘᵢₙ make sense.

SphericalFunctions.Yiterator — Type

Yiterator(Y, ℓₘₐₓ, [ℓₘᵢₙ, [iₘᵢₙ]])

Construct an Iterator that returns sub-vectors of Y, each of which consists of elements $(ℓ,-ℓ)$ through $(ℓ,ℓ)$, for $ℓ$ from ℓₘᵢₙ through ℓₘₐₓ.

Note that the returned objects are views into the original Y data — meaning that you may alter their values.

Because the result is a vector restricted to a particular $ℓ$ value, you can index the $(ℓ, m)$ element as [ℓ+m+1]. For example, you might use this as something like

for (ℓ, Yˡ) in zip(ℓₘᵢₙ:ℓₘₐₓ, Yiterator(Y, ℓₘₐₓ))
    for m in -ℓ:ℓ
        Yˡ[ℓ+m+1] = <...>
    end
end

By default, Y is assumed to contain all possible values, beginning with (0,0). However, if ℓₘᵢₙ is not 0, this can be ambiguous: do we mean that Y really starts with the (0,0) element and we are just asking to begin the iteration higher? Or do we mean that Y doesn't even contain data for lower ℓ values? We can resolve this using iₘᵢₙ, which gives the index of ℓₘᵢₙ in Y. By default, we assume the first case, and set iₘᵢₙ=Ysize(ℓₘᵢₙ-1)+1. However, if Y doesn't contain data below ℓₘᵢₙ, we could use iₘᵢₙ=1 to indicate the index in Y at which to find $(ℓₘᵢₙ,-ℓₘᵢₙ)$.

Also note that no bounds checking is done, either at instantiation time or during iteration. You are responsible for ensuring that the size of Y and the values of ℓₘₐₓ, ℓₘᵢₙ, and iₘᵢₙ make sense.

SphericalFunctions.diterator — Type

diterator(d, ℓₘₐₓ, [ℓₘᵢₙ])

Construct an Iterator that returns sub-matrices of d, each of which consists of elements $(ℓ,-ℓ,-ℓ)$ through $(ℓ,ℓ,ℓ)$, for $ℓ$ from ℓₘᵢₙ through ℓₘₐₓ. By default, ℓₘᵢₙ is 0.

Note that the returned objects are views into the original d data — meaning that you may alter their values.

Because the result is a matrix restricted to a particular $ℓ$ value, you can index the $(ℓ, m′, m)$ element as [ℓ+m′+1, ℓ+m+1]. For example, you might use this as something like

for (ℓ, dˡ) in zip(ℓₘᵢₙ:ℓₘₐₓ, diterator(d, ℓₘₐₓ))
    for m′ in -ℓ:ℓ
        for m in -ℓ:ℓ
            dˡ[ℓ+m′+1, ℓ+m+1] = <...>
        end
    end
end

Also note that no bounds checking is done, either at instantiation time or during iteration. You are responsible for ensuring that the size of d and the values of ℓₘₐₓ and ℓₘᵢₙ make sense.

Transformation

The functions above — especially Y! — allow us to evaluate mode weights (often denoted salm) of a spin-weighted spherical-harmonic expansion as a function of a particular point (often denoted map). This process is frequently referred to as "synthesis", and is also called salm2map.

We can also perform the reverse process of going from the function evaluated on some kind of grid, back to the corresponding mode weights — also called "analysis" or map2salm.

SphericalFunctions.map2salm! — Method

map2salm!(salm, map, spin, ℓmax)
map2salm!(salm, map, plan)

Transform map values sampled on the sphere to ${}_sa_{\ell, m}$ modes in place.

For details, see map2salm.

SphericalFunctions.map2salm — Method

map2salm(map, spin, ℓmax)
map2salm(map, plan)

Transform map values sampled on the sphere to ${}_sa_{\ell, m}$ modes.

The map array should have size Nφ along its first dimension and Nϑ along its second; any number of dimensions may follow. The spin must be entered explicitly, and ℓmax is the highest ℓ value you want in the output.

For repeated applications of this function with different values of map, it is more efficient to pre-compute plan using plan_map2salm. These functions will create a new salm array on each call. To operate in place on a pre-allocated salm array, use map2salm!.

The core of this function follows the method described by Reinecke and Seljebotn.

SphericalFunctions.plan_map2salm — Method

plan_map2salm(map, spin, ℓmax)

Precompute values to use in executing map2salm or map2salm!.

The arguments to this function exactly mirror those of the first form of map2salm, and all but the first argument in the first form of map2salm!. The plan returned by this function can be passed to the second forms of those functions to avoid some computation and allocation costs.

Note that the plan object is not thread safe; a separate plan should be created for each thread that will use one, or locks should be used to ensure that a single plan is not used at the same time on different threads.