Function calculators

Typically, when calculating special functions, we will use recursion relations along with some coefficients — which frequently requires significant setup processing. That processing can be cached, so that the calculations themselves consist primarily of memory accesses and simple arithmetic.

Similarly, some computations require a certain amount of "workspace" — significant amounts of memory that are needed for the computation, but are not part of the core result (and don't even need to be returned). Rather than allocating this memory on each call to the computation, we can allocate once and pass the workspace in as an argument.

These two patterns — pre-computed coefficients and pre-allocated workspaces — will appear as possible call signatures in each of the following functions. Note that we also have helper functions like Dprep, Dstorage, etc., which construct the necessary coefficients and/or workspaces, as well as the actual results for in-place computations, to make it easier to call the corresponding calculator functions.

Note that these functions impose assumptions about the ordering of arrays as vectors. For example, the series of Wigner $𝔇$ matrices will be stored as a single vector. And for particular values of $(ℓ, n, m)$, the element $𝔇ˡₙ,ₘ$ will have to be accessed as some linearly indexed element of that vector. These formats, along with utility functions for accessing particular elements, are described here.

The fundamental algorithm is the $H$ recursion, which is the core computation needed for Wigner's $d$ and $𝔇$ matrices, and the spin-weighted spherical harmonics ${}_{s}Y_{\ell,m}$, as well as map2salm functions.

SphericalFunctions.H! — Method

H!(H, expiβ, ℓₘₐₓ, m′ₘₐₓ, H_rec_coeffs)
H!(H, expiβ, ℓₘₐₓ, m′ₘₐₓ, H_rec_coeffs, Hindex)

Compute the $H$ matrix defined by Gumerov and Duraiswami.

This computation forms the basis for computing Wigner's $d$ and $𝔇$ matrices via d! and D!, the spin-weighted spherical harmonics via Y!, and for transforming from values of spin-weighted spherical functions evaluated on a grid to the corresponding mode weights via map2salm.

Due to symmetries, we only need to compute ~1/4 of the elements of this matrix, so only those elements with $m ≥ |m′|$ are computed. The relevant indices of the H vector are computed based on the Hindex function — which defaults to WignerHindex, but could reasonably be WignerDindex if the input H vector contains all valid indices. However, it is assumed that the storage scheme used for H is such that the successive $m$ values are located in successive elements.

If $m′ₘₐₓ < ℓₘₐₓ$, we don't even need 1/4 of the elements, and only values with $|m′| ≤ m′ₘₐₓ$ will be computed. This is particularly useful for computing spin-weighted spherical harmonics.

Note that the recursion coefficients H_rec_coeffs should be the quantity returned by H_recursion_coefficients.

SphericalFunctions.H_recursion_coefficients — Method

H_recursion_coefficients(ℓₘₐₓ, T)

Pre-compute constants used in Wigner H recursion.

Internally, the $H$ recursion relies on calculation of the Associated Legendre Functions (ALFs), which can also be called on their own:

SphericalFunctions.ALFcompute! — Method

ALFcompute(expiβ, nmax)
ALFcompute!(P̄, expiβ, nmax)
ALFcompute(expiβ, nmax, recursion_coefficients)
ALFcompute!(P̄, expiβ, nmax, recursion_coefficients)

Compute the "fully normalized" Associated Legendre Functions up to some maximum n value nmax.

These functions can take a vector P̄, to store the data, stored in order of increasing m most rapidly varying and then increasing n. If not supplied, P̄ will be constructed for you and returned.

The optional recursion_coefficients argument must be an ALFRecursionCoefficients, which stores various constant coefficients used in the recursion. This object requires more than 3x the memory and more than 20x the time to compute a single P̄ vector without this argument, but passing it will typically speed up the calculation of each P̄ by a factor of 8x or so. Thus, if you expect to compute P̄ more than a few times, it will take less time to pre-compute those factors, and pass them to this function.

Note that the base real type will be inferred from the (complex) type of expiβ. If present, the base types of P̄ and recursion_coefficients must agree.

Based on those, we have the d!, D!, and Y! functions:

SphericalFunctions.D! — Method

D!(𝔇, R, ℓₘₐₓ, (aₙᵐ,bₙᵐ,dₙᵐ), eⁱᵐᵅ, eⁱᵐᵞ)

Compute Wigner's 𝔇 matrix

This function implements the preferred method of computing the 𝔇 matrix for large ell values. In particular, above ell≈32 standard formulas become completely unusable because of numerical instabilities and overflow. This function uses stable recursion methods instead, and should be usable beyond ell≈1000.

This function computes 𝔇ˡₘₚ,ₘ(R). The result is returned in a 1-dimensional array ordered as

[
    𝔇(ell, mp, m, R)
    for ell in range(ell_max+1)
    for mp in range(-min(ℓ, mp_max), min(ℓ, mp_max)+1)
    for m in range(-ell, ell+1)
]

SphericalFunctions.Dprep — Method

Dprep(ℓₘₐₓ, T)

Construct space and pre-compute recursion coefficients to compute Wigner's $𝔇$ matrix in place.

This returns the (D, H_rec_coeffs, eⁱᵐᵅ, eⁱᵐᵞ) arguments needed by D!.

SphericalFunctions.Dstorage — Method

Dstorage(ℓₘₐₓ, T)

Construct space to compute Wigner's $𝔇$ matrix in place.

This returns the D argument needed by D!.

SphericalFunctions.Y! — Function

Y!(Y, R, ℓₘₐₓ, spin, H_rec_coeffs, Hwedge, expimϕ, ℓₘᵢₙ=0)
Y!(Y, expiθ, expiϕ, ℓₘₐₓ, spin, H_rec_coeffs, Hwedge, expimϕ, ℓₘᵢₙ=0)
Y!(Y, expiϕ, expiθ, expiγ, ℓₘₐₓ, spin, H_rec_coeffs, Hwedge, expimϕ, ℓₘᵢₙ=0)

Evaluate (and write into Y, if present) the values of ${}_{s}Y_{\ell, m}(R)$ for the input value of s, for all $(\ell, m)$ throughout the range specified by wigner. R is assumed to be a unit quaternion (which may be Rotor, or simply a Quaternion). If R does not have unit magnitude, the output elements will be too large by a factor $|R|^{\ell}$. If Y is not present, a new array will be created.

The spherical harmonics of spin weight $s$ are related to Wigner's $\mathfrak{D}$ matrix as

\[\begin{aligned} {}_{s}Y_{\ell, m}(R) &= (-1)^s \sqrt{\frac{2\ell+1}{4\pi}} \mathfrak{D}^{(\ell)}_{m, -s}(R) \\ &= (-1)^s \sqrt{\frac{2\ell+1}{4\pi}} \bar{\mathfrak{D}}^{(\ell)}_{-s, m}(\bar{R}). \end{aligned}\]

SphericalFunctions.Yprep — Method

Yprep(ℓₘₐₓ, sₘₐₓ, T, ℓₘᵢₙ)

Prepare the storage, recursion coefficients, and workspace to compute ₛYₗ,ₘ data up to the maximum sizes given.

Returns a tuple of Y, H_rec_coeffs, Hwedge, expimϕ, which can be passed to the correspondingly named arguments of Y!.

Note that the same results of this function can be passed to Y!, even if the value of ℓₘₐₓ passed to that function is smaller than the value passed to this function, or the value of spin passed to that function is smaller (in absolute value) than the sₘₐₓ passed to this function. However, the value of ℓₘᵢₙ passed to that function must not be smaller than the value passed to this function (unless one of the other sizes is sufficiently smaller).

SphericalFunctions.Ystorage — Method

Ystorage(ℓₘₐₓ, T)
Ystorage(ℓₘₐₓ, T, ℓₘᵢₙ)

Construct space to compute the spin-weighted spherical harmonics $ₛYₗ,ₘ$ in place.

This returns the Y argument needed by Y!.

SphericalFunctions.d! — Method

d!(d, expiβ, ℓₘₐₓ, H_rec_coeffs)
d!(d, expiβ, ℓₘₐₓ)
d!(d, β, ℓₘₐₓ, H_rec_coeffs)
d!(d, β, ℓₘₐₓ)
d(expiβ, ℓₘₐₓ)
d(β, ℓₘₐₓ)

Compute Wigner's d matrix dˡₘₚ,ₘ(β)

Notes

This function is the preferred method of computing the d matrix for large ell values. In particular, above ell≈32 standard formulas become completely unusable because of numerical instabilities and overflow. This function uses stable recursion methods instead, and should be usable beyond ell≈1000.

The result is returned in a 1-dimensional array ordered as

[
    d(ell, mp, m, β)
    for ell in range(ell_max+1)
    for mp in range(-min(ℓ, mp_max), min(ℓ, mp_max)+1)
    for m in range(-ell, ell+1)
]

SphericalFunctions.dprep — Method

dprep(ℓₘₐₓ, T)

Construct space and pre-compute recursion coefficients to compute Wigner's $d$ matrix in place.

This returns the (d, H_rec_coeffs) arguments needed by d!.

SphericalFunctions.dstorage — Method

dstorage(ℓₘₐₓ, T)

Construct space to compute Wigner's $d$ matrix in place.

This returns the d argument needed by d!.

SphericalFunctions.ₛ𝐘 — Method

ₛ𝐘(s, ℓₘₐₓ, [T=Float64], [Rθϕ=golden_ratio_spiral_rotors(s, ℓₘₐₓ, T)])

Construct a matrix of $ₛYₗₘ(Rθϕ)$ values for the input s and all nontrivial $(\ell, m)$ up to ℓₘₐₓ.

This is a fast and accurate method for mapping between the vector of spin-weighted spherical-harmonic mode weights $ₛ𝐟ₗₘ$ and the vector of function values on the sphere $ₛ𝐟ⱼₖ$, as

\[ₛ𝐟ⱼₖ = ₛ𝐘\, ₛ𝐟ₗₘ,\]

where the right-hand side represents the matrix-vector product. As usual, we assume that the $ₛ𝐟ₗₘ$ modes are ordered by increasing $m ∈ [-ℓ:ℓ]$, and $ℓ ∈ [|s|:ℓₘₐₓ]$. The ordering of the $ₛ𝐟ⱼₖ$ values will be determined by the ordering of the argument Rθϕ.

Note that the number of modes need not be the same as the number of points on which the function is evaluated, which would imply that the output matrix is not square. To be able to invert the relationship, however, we need the number of points $ₛ𝐟ⱼₖ$ to be at least as large as the number of modes $ₛ𝐟ₗₘ$.

Note that the usefulness of this approach is limited by the fact that the size of this matrix scales as ℓₘₐₓ⁴. As such, it is mostly useful only for ℓₘₐₓ of order dozens, rather than — say — the tens of thousands that CMB astronomy or lensing require, for example.

Direct application and inversion of this matrix are used in the "direct" methods of $s$-SHT transformations. See SSHTDirect for details about the implementation.

In all cases, the returned (or overwritten) data are stored linearly, as a single Vector ranging over all indices. To access subarrays based on certain $ℓ$ values, for example, we also have iterators that return conveniently shaped views into these linear vectors:

SphericalFunctions.AlternatingCountdown — Type

Simple iterator to count down to 0, with alternating signs

julia> collect(AlternatingCountdown(5))
11-element Vector{Int64}:
  5
 -5
  4
 -4
  3
 -3
  2
 -2
  1
 -1
  0

SphericalFunctions.Diterator — Type

Diterator(D, ℓₘₐₓ, [ℓₘᵢₙ])

Construct an Iterator that returns sub-matrices of D, each of which consists of elements $(ℓ,-ℓ,-ℓ)$ through $(ℓ,ℓ,ℓ)$, for $ℓ$ from ℓₘᵢₙ through ℓₘₐₓ. By default, ℓₘᵢₙ is 0.

Note that the returned objects are views into the original D data — meaning that you may alter their values.

Because the result is a matrix restricted to a particular $ℓ$ value, you can index the $(ℓ, m′, m)$ element as [ℓ+m′+1, ℓ+m+1]. For example, you might use this as something like

for (ℓ, Dˡ) in zip(ℓₘᵢₙ:ℓₘₐₓ, Diterator(D, ℓₘₐₓ))
    for m′ in -ℓ:ℓ
        for m in -ℓ:ℓ
            Dˡ[ℓ+m′+1, ℓ+m+1] = <...>
        end
    end
end

Also note that no bounds checking is done, either at instantiation time or during iteration. You are responsible for ensuring that the size of D and the values of ℓₘₐₓ and ℓₘᵢₙ make sense.

SphericalFunctions.Yiterator — Type

Yiterator(Y, ℓₘₐₓ, [ℓₘᵢₙ, [iₘᵢₙ]])

Construct an Iterator that returns sub-vectors of Y, each of which consists of elements $(ℓ,-ℓ)$ through $(ℓ,ℓ)$, for $ℓ$ from ℓₘᵢₙ through ℓₘₐₓ.

Note that the returned objects are views into the original Y data — meaning that you may alter their values.

Because the result is a vector restricted to a particular $ℓ$ value, you can index the $(ℓ, m)$ element as [ℓ+m+1]. For example, you might use this as something like

for (ℓ, Yˡ) in zip(ℓₘᵢₙ:ℓₘₐₓ, Yiterator(Y, ℓₘₐₓ))
    for m in -ℓ:ℓ
        Yˡ[ℓ+m+1] = <...>
    end
end

By default, Y is assumed to contain all possible values, beginning with (0,0). However, if ℓₘᵢₙ is not 0, this can be ambiguous: do we mean that Y really starts with the (0,0) element and we are just asking to begin the iteration higher? Or do we mean that Y doesn't even contain data for lower ℓ values? We can resolve this using iₘᵢₙ, which gives the index of ℓₘᵢₙ in Y. By default, we assume the first case, and set iₘᵢₙ=Ysize(ℓₘᵢₙ-1)+1. However, if Y doesn't contain data below ℓₘᵢₙ, we could use iₘᵢₙ=1 to indicate the index in Y at which to find $(ℓₘᵢₙ,-ℓₘᵢₙ)$.

Also note that no bounds checking is done, either at instantiation time or during iteration. You are responsible for ensuring that the size of Y and the values of ℓₘₐₓ, ℓₘᵢₙ, and iₘᵢₙ make sense.

SphericalFunctions.diterator — Type

diterator(d, ℓₘₐₓ, [ℓₘᵢₙ])

Construct an Iterator that returns sub-matrices of d, each of which consists of elements $(ℓ,-ℓ,-ℓ)$ through $(ℓ,ℓ,ℓ)$, for $ℓ$ from ℓₘᵢₙ through ℓₘₐₓ. By default, ℓₘᵢₙ is 0.

Note that the returned objects are views into the original d data — meaning that you may alter their values.

Because the result is a matrix restricted to a particular $ℓ$ value, you can index the $(ℓ, m′, m)$ element as [ℓ+m′+1, ℓ+m+1]. For example, you might use this as something like

for (ℓ, dˡ) in zip(ℓₘᵢₙ:ℓₘₐₓ, diterator(d, ℓₘₐₓ))
    for m′ in -ℓ:ℓ
        for m in -ℓ:ℓ
            dˡ[ℓ+m′+1, ℓ+m+1] = <...>
        end
    end
end

Also note that no bounds checking is done, either at instantiation time or during iteration. You are responsible for ensuring that the size of d and the values of ℓₘₐₓ and ℓₘᵢₙ make sense.

SphericalFunctions.λiterator — Type

λiterator(θ, s, m)

Construct an object to iterate over ₛλₗₘ values.

The $ₛλₗₘ(θ)$ function is defined as the spin-weighted spherical harmonic evaluated at spherical coordinates $(θ, ϕ)$, with $ϕ=0$. In particular, note that it is real-valued. The return type is determined by the type of θ (or more precisely, cos½θ).

This algorithm by Kostelec & Rockmore allows us to iterate over increasing $ℓ$ values, for given fixed $s$ and $m$ values.

Note that this iteration has no inherent bound, so if you try to iterate over all values, you will end up in an infinite loop. Instead, you can zip this iterator with another:

θ = 0.1
s = -2
m = 1
λ = λiterator(θ, s, m)
Δ = max(abs(s), abs(m))
for (ℓ, ₛλₗₘ) ∈ zip(Δ:Δ+5, λ)
    @show (ℓ, ₛλₗₘ)
end

Alternatively, you could use Iterates.take(λ, 6), for example.

Note that the iteration always begins with ℓ = Δ = max(abs(s), abs(m)).

SphericalFunctions.λ_recursion_initialize — Method

λ_recursion_initialize(cosθ, sin½θ, cos½θ, s, ℓ, m)

This provides initial values for the recursion to find ${}_{s}\lambda_{\ell,m}$ along indices of increasing $\ell$, due to Kostelec & Rockmore. Specifically, this function computes values with $\ell=m$.

\[{}_{s}\lambda_{\ell,m}(\theta) := {}_{s}Y_{\ell,m}(\theta, 0) = (-1)^m\, \sqrt{\frac{2\ell+1}{4\pi}} d^\ell_{-m,s}(\theta)\]

Transformation

The functions above — especially Y! — allow us to evaluate mode weights (often denoted salm) of a spin-weighted spherical-harmonic expansion as a function of a particular point (often denoted map). This process is frequently referred to as "synthesis", and is also called salm2map.

We can also perform the reverse process of going from the function evaluated on some kind of grid, back to the corresponding mode weights — also called "analysis" or map2salm.

SphericalFunctions.map2salm! — Method

map2salm!(salm, map, spin, ℓmax)
map2salm!(salm, map, plan)

Transform map values sampled on the sphere to ${}_sa_{\ell, m}$ modes in place.

For details, see map2salm.

SphericalFunctions.map2salm — Method

map2salm(map, spin, ℓmax)
map2salm(map, plan)

Transform map values sampled on the sphere to ${}_sa_{\ell, m}$ modes.

The map array should have size Nφ along its first dimension and Nϑ along its second; any number of dimensions may follow. The spin must be entered explicitly, and ℓmax is the highest ℓ value you want in the output.

For repeated applications of this function with different values of map, it is more efficient to pre-compute plan using plan_map2salm. These functions will create a new salm array on each call. To operate in place on a pre-allocated salm array, use map2salm!.

The core of this function follows the method described by Reinecke and Seljebotn.

SphericalFunctions.plan_map2salm — Method

plan_map2salm(map, spin, ℓmax)

Precompute values to use in executing map2salm or map2salm!.

The arguments to this function exactly mirror those of the first form of map2salm, and all but the first argument in the first form of map2salm!. The plan returned by this function can be passed to the second forms of those functions to avoid some computation and allocation costs.

Note that the plan object is not thread safe; a separate plan should be created for each thread that will use one, or locks should be used to ensure that a single plan is not used at the same time on different threads.

A newer interface allows simpler syntax, akin to that of FFTW.jl, whereby a transformation object 𝒯 can be used to transform between function values 𝐟 and mode weights 𝐦 as either

𝐟 = 𝒯 * 𝐦

or

𝐦 = 𝒯 \ 𝐟

SphericalFunctions.SSHT — Type

Supertype of storage for spin-spherical-harmonic transforms

SphericalFunctions.SSHT — Method

SSHT(s, ℓₘₐₓ; [method="Direct"], [T=Float64], [kwargs...])

Construct an SSHT object to transform between spin-weighted spherical-harmonic mode weights and function values — performing an $s$-SHT.

This object behaves similarly to an AbstractFFTs.Plan object — specfically in the ability to use the semantics of algebra to perform transforms. For example, if the function values are stored as a vector f, the mode weights as f̃, and the SSHT as 𝒯, then we can compute the function values from the mode weights as

f = 𝒯 * f̃

or solve for the mode weights from the function values as

f̃ = 𝒯 \ f

The first dimensions of f̃ must index the mode weights (as usual, for ℓ∈abs(m):ℓₘₐₓ and m∈-ℓ:ℓ) and the first index of f must index the locations at which the function is evaluated. Any following dimensions will be broadcast over. Note that certain types will broadcast using Julia threads, while others will broadcast using BLAS threads. The relevant number of threads must be set appropriately.

Certain SSHT types (currently, only Minimal and Direct) also have an option to always act in place — meaning that they simply re-use the input storage, even when used in an expression like 𝒯 \ f. The option must be passed as the inplace argument to the constructors, and part of the type of the resulting object. Regardless of the value of that option, for those types where the option exists, it is also possible to use mul! and ldiv! from the LinearAlgebra package to force operation in place.

SphericalFunctions.pixels — Function

pixels(𝒯)

Return the spherical coordinates (θ, ϕ) on which the spin-weighted spherical harmonics are evaluated. See also rotors, which provides the actual Rotors on which they are evaluated.

SphericalFunctions.rotors — Function

rotors(𝒯)

Return the Rotors on which the spin-weighted spherical harmonics are evaluated. See also pixels, which provides the corresponding spherical coordinates.

SphericalFunctions.SSHTDirect — Type

SSHTDirect(s, ℓₘₐₓ; decomposition=LinearAlgebra.qr, T=Float64, Rθϕ=golden_ratio_spiral_rotors(s, ℓₘₐₓ, T), inplace=true)

Construct an $s$-SHT object that uses the "Direct" method; see ₛ𝐘 for details about the method and optional arguments. Also see SSHT for general information about how to use these objects.

By default, this uses precisely optimal sampling — meaning that the number of points on which the function is evaluated, represented by Rθϕ, is equal to the number of modes. However, it is equally possible to evaluate on more points than there are modes. This can be useful, for example, when processing multiple fields with different spin weights; the function could be evaluated on points appropriate for the lowest value of $|s|$, and therefore could also be used to solve for fields of all other spin weights.

Note that in-place operation is possible for this type when the length of the input Rθϕ is equal to the number of modes given s and ℓₘₐₓ — and is the default behavior when possible. See SSHT for description of in-place operation.

This method is typically better than other current implementations for $ℓₘₐₓ ≲ 24$, both in terms of speed and accuracy. However, this advantage quickly falls away. A warning will be issued if ℓₘₐₓ is greater than about 64, because this method is not likely to be the most efficient or most accurate choice.

SphericalFunctions.SSHTMinimal — Type

Storage for Minimal spin-spherical-harmonic transform

The Minimal algorithm was described in this paper, and allows for the minimal number of function samples.

SphericalFunctions.SSHTRS — Type

Storage for spin-spherical-harmonic transform

The algorithm was described in this paper by Reinecke and Seljebotn.

SphericalFunctions.SSHTRS — Method

SSHTRS(s, ℓₘₐₓ; kwargs...)

Construct a SSHTRS object directly. This may also be achieved by calling the main SSHT function with the same keywords, along with method="RS".

This object uses the algorithm described in this paper by Reinecke and Seljebotn.

The basic floating-point number type may be adjusted with the keyword argument T, which defaults to Float64.

The SSHs are evaluated on a series of "rings" at constant colatitude. Their locations are specified by the θ keyword argument, which defaults to fejer1_rings(2ℓₘₐₓ+1, T). If this is changed, the user should also provide the corresponding quadrature_weights argument — the default being fejer1(length(θ), T).

On each of these rings, an FFT is performed. To reach the band limit of $m = ± ℓₘₐₓ$, the number of points along each ring must therefore be at least $2ℓₘₐₓ+1$, but may be greater. For example, if $2ℓₘₐₓ+1$ does not factorize neatly into a product of small primes, it may be preferable to use $2ℓₘₐₓ+2$ points along each ring. (In that case, whenever ℓₘₐₓ is 1 less than a power of 2, the number of points will be exactly a power of 2, which is usually particularly efficient.) The number of points on each ring can be modified independently, if given as a vector with the same length as θ, or as a single number which is assumed to be the same for all rings.

Whenever T is either Float64 or Float32, the keyword arguments plan_fft_flags and plan_fft_timelimit may also be useful for obtaining more efficient FFTs. They default to FFTW.ESTIMATE and Inf, respectively. They are passed to AbstractFFTs.plan_fft.

The SSHT types use various pixelizations that may be computed as follows:

SphericalFunctions.clenshaw_curtis_rings — Method

clenshaw_curtis_rings(N, [T=Float64])

Values of the colatitude coordinate ($θ$) appropriate for quadrature by the Clenshaw-Curtis rule, using weights provided by clenshaw_curtis.

Note that the first argument to this function is N, rather than the ℓₘₐₓ used in some other functions. For spin-weighted spherical harmonics, you may want to use N=2ℓₘₐₓ+1.

SphericalFunctions.fejer1_rings — Method

fejer1_rings(N, [T=Float64])

Values of the colatitude coordinate ($θ$) appropriate for quadrature by Fejér's first rule, using weights provided by fejer1.

Note that the first argument to this function is N, rather than the ℓₘₐₓ used in some other functions. For spin-weighted spherical harmonics, you may want to use N=2ℓₘₐₓ+1.

SphericalFunctions.fejer2_rings — Method

fejer2_rings(N, [T=Float64])

Values of the colatitude coordinate ($θ$) appropriate for quadrature by Fejér's second rule, using weights provided by fejer2.

Note that the first argument to this function is N, rather than the ℓₘₐₓ used in some other functions. For spin-weighted spherical harmonics, you may want to use N=2ℓₘₐₓ+1.

SphericalFunctions.golden_ratio_spiral_pixels — Method

golden_ratio_spiral_pixels(s, ℓₘₐₓ, [T=Float64])

Cover the sphere 𝕊² with pixels generated by the golden-ratio spiral. Successive pixels are separated by the azimuthal angle $Δϕ = 2π(2-φ)$, and are uniformly distributed in $\cos θ$.

Visually, this is a very reasonable-looking pixelization, with fairly uniform distance between neighbors, and approximate isotropy. No two pixels will share the same values of either $θ$ or $ϕ$. Also note that no point is present on either the North or South poles.

The returned quantity is a vector of 2-SVectors providing the spherical coordinates of each pixel. See also golden_ratio_spiral_rotors for the corresponding Rotors.

SphericalFunctions.golden_ratio_spiral_rotors — Method

golden_ratio_spiral_rotors(s, ℓₘₐₓ, [T=Float64])

Cover the sphere 𝕊² with pixels generated by the golden-ratio spiral.

See golden_ratio_spiral_pixels for more detailed explanation. The quantity returned by this function is a vector of Rotors providing each pixel.

SphericalFunctions.sorted_ring_pixels — Method

sorted_ring_pixels(s, ℓₘₐₓ, [T=Float64])

Cover the sphere 𝕊² with $(ℓₘₐₓ+1)²-s²$ pixels distributed in rings provided by sorted_rings; see that function's documentation for more description.

The returned quantity is a vector of 2-SVectors containing the spherical coordinates of each pixel. See also sorted_ring_rotors for the corresponding Rotors.

SphericalFunctions.sorted_ring_rotors — Method

sorted_ring_rotors(s, ℓₘₐₓ, [T=Float64])

Cover the sphere 𝕊² with $(ℓₘₐₓ+1)²-s²$ pixels distributed in rings provided by sorted_rings; see that function's documentation for more description.

The returned quantity is a vector of Rotors. See also sorted_ring_rotors for the corresponding spherical coordinates.

SphericalFunctions.sorted_rings — Method

sorted_rings(s, ℓₘₐₓ, [T=Float64])

Compute locations of a series of rings labelled by $j ∈ |s|:ℓₘₐₓ$ (analogous to $ℓ$), where each ring will contain $k = 2j+1$ (analogous to $m$) pixels distributed evenly around the ring. These rings are then sorted, so that the ring with the most pixels ($j = ℓₘₐₓ$) is closest to the equator, and the next-largest ring is placed just above or below the equator (depending on the sign of $s$), the next just below or above, and so on. This is generally a fairly good first guess when minimizing the condition number of matrices used to solve for mode weights from function values. In particular, I use this to initialize the Minimal algorithm, which is then fed into an optimizer to fine-tune the positions of the rings.

This function does not provide the individual pixels; it just provides the colatitude values of the rings on which the pixels will be placed. The pixels themselves are provided by sorted_ring_pixels.