Utilities

While not usually the star of the show, the following utilities can be quite helpful for actually using the rest of the code.

Complex powers

One common task we find when working with spherical functions is the computation of a range of integer powers of some complex number — so much so that it can be best to pre-compute the powers and cache their values. While a naive approach is generally quite accurate, and reasonably fast, we can do a little better with a specialized routine.

SphericalFunctions.complex_powers! — Method

complex_powers!(zpowers, z)

Compute integer powers of z from z^0 through z^m, recursively, where m is one less than the length of the input zpowers vector.

Note that z is assumed to be normalized, with complex amplitude approximately 1.

See also: complex_powers

SphericalFunctions.complex_powers — Method

complex_powers(z, m)

Compute integer powers of z from z^0 through z^m, recursively.

Note that z is assumed to be normalized, with complex amplitude approximately 1.

This algorithm is mostly due to Stoer and Bulirsch in "Introduction to Numerical Analysis" (page 24) — with a little help from de Moivre's formula, which is essentially exp(iθ)ⁿ = exp(inθ), as well as my own alterations to deal with different behaviors in different quadrants.

There isn't usually a huge advantage to using this specialized function. If you just need a particular power, it will generally be far more efficient and just as accurate to compute either exp(iθ)ⁿ or exp(inθ) explicitly. However, if you need all powers from 0 to m, this function is about 10 or 5 times faster than those options, respectively, for large m. Like those options, this function is numerically stable, in the sense that its errors are usually smaller than m times the error from machine-precision errors in the input argument — or at worst about 50% larger, which occurs as the phase approaches multiples of π/2.

See also: complex_powers!

Quadrature weights

SphericalFunctions.clenshaw_curtis — Method

clenshaw_curtis(n, [T])

Compute n weights for the Clenshaw-Curtis rule, corresponding to n evenly spaced nodes from 0 to π inclusive. That is, the nodes are located at

\[\theta_k = k \frac{\pi}{n-1} \quad k=0, \ldots, n-1.\]

This function uses Waldvogel's method.

The type T may be any AbstractFloat, but defaults to Float64.

SphericalFunctions.fejer1 — Method

fejer1(n, [T])

Compute n weights for Fejér's first rule, corresponding to n evenly spaced nodes from 0 to π inclusive. That is, the nodes are located at

\[\theta_k = k \frac{\pi}{n-1} \quad k=0, \ldots, n-1.\]

This function uses Waldvogel's method.

The type T may be any AbstractFloat, but defaults to Float64.

SphericalFunctions.fejer2 — Method

fejer2(n, [T])

Compute n weights for Fejér's second rule, corresponding to n evenly spaced nodes between 0 and π exclusive. That is, the nodes are located at

\[\theta_k = k \frac{\pi}{n+1} \quad k=1, \ldots, n.\]

This function uses Waldvogel's method. However, contrary to Waldvogel's notation, this routine does not include the weight corresponding to the ϑ=0 or π nodes, which both have weight 0.

The type T may be any AbstractFloat, but defaults to Float64.

$Y$ and $D$ data

By $Y$ data, we mean anything indexed like $Y_{\ell, m}$ modes; by $D$ data, we mean anything indexed like Wigner's $\mathfrak{D}$ matrices, or special subsets of them, like the $H$ matrices.

SphericalFunctions.WignerDindex — Function

WignerDindex(ℓ, m′, m, m′ₘₐₓ=ℓ)

Compute index into Wigner 𝔇 matrix

Parameters

ℓ : int Integer satisfying ℓₘᵢₙ <= ℓ <= ℓₘₐₓ m′ : int Integer satisfying -min(ℓₘₐₓ, m′ₘₐₓ) <= m′ <= min(ℓₘₐₓ, m′ₘₐₓ) m : int Integer satisfying -ℓ <= m <= ℓ ℓₘᵢₙ : int, optional Integer satisfying 0 <= ℓₘᵢₙ <= ℓₘₐₓ. Defaults to 0. m′ₘₐₓ : int, optional Integer satisfying 0 <= m′ₘₐₓ. Defaults to ℓ.

Returns

i : int Index into Wigner 𝔇 matrix arranged as described below

See Also

WignerDsize : Total size of the 𝔇 matrix WignerDrange : Array of (ℓ, m', m) indices corresponding to the 𝔇 matrix

Notes

This assumes that the Wigner 𝔇 matrix is arranged as

[
    𝔇(ℓ, m′, m)
    for ℓ in range(ℓₘᵢₙ, ℓₘₐₓ+1)
    for m′ in range(-min(ℓ, m′ₘₐₓ), min(ℓ, m′ₘₐₓ)+1)
    for m in range(-ℓ, ℓ+1)
]

SphericalFunctions.WignerDrange — Function

WignerDrange(ℓₘₐₓ, m′ₘₐₓ=ℓₘₐₓ)

Create an array of (ℓ, m', m) indices as in 𝔇 array

Parameters

ℓₘᵢₙ : int Integer satisfying 0 <= ℓₘᵢₙ <= ℓₘₐₓ m′ₘₐₓ : int, optional Integer satisfying 0 <= m′ₘₐₓ. Default is ℓₘₐₓ. ℓₘₐₓ : int Integer satisfying 0 <= ℓₘᵢₙ <= ℓₘₐₓ

See Also

WignerDsize : Total size of 𝔇 array WignerDindex : Index inside these wedges

Notes

This assumes that the Wigner 𝔇 matrix is arranged as

[
    𝔇(ℓ, m′, m)
    for ℓ in range(ℓₘᵢₙ, ℓₘₐₓ+1)
    for m′ in range(-min(ℓ, m′ₘₐₓ), min(ℓ, m′ₘₐₓ)+1)
    for m in range(-ℓ, ℓ+1)
]

SphericalFunctions.WignerDsize — Method

WignerDsize(ℓₘₐₓ, m′ₘₐₓ=ℓₘₐₓ)

Compute total size of Wigner 𝔇 matrix

See also 'WignerDrange` and WignerDindex.

Notes

This assumes that the Wigner 𝔇 matrix is arranged as

[
    𝔇(ℓ, m′, m)
    for ℓ in ℓₘᵢₙ:ℓₘₐₓ
    for m′ in -min(ℓ, m′ₘₐₓ):min(ℓ, m′ₘₐₓ)
    for m in -ℓ:ℓ
]

SphericalFunctions.WignerHindex — Function

WignerHindex(ℓ, m′, m, m′ₘₐₓ)

Index to "wedge" arrays

Parameters

ℓ : int m′ : int m : int m′ₘₐₓ : int, optional If nothing, it is assumed to be at least ℓ

See Also

WignerHsize : Total size of wedge array WignerHrange : Array of (ℓ, m', m) indices corresponding to this wedge

Notes

Here, it is assumed that only data with m≥|m'| are stored, and only corresponding values are passed. We also assume |m|≤ℓ and |m'|≤ℓ. Neither of these are checked. The wedge array that this function indexes is ordered as

[
    H(ℓ, m′, m) for ℓ in range(ℓₘₐₓ+1)
    for m′ in range(-min(ℓ, m′ₘₐₓ), min(ℓ, m′ₘₐₓ)+1)
    for m in range(abs(m′), ℓ+1)
]

SphericalFunctions.WignerHrange — Function

WignerHrange(ℓₘₐₓ, m′ₘₐₓ=ℓₘₐₓ)

Create an array of (ℓ, m', m) indices as in H array

See also 'WignerHsize` and WignerHindex

Notes

Here, it is assumed that only data with m≥|m'| are stored, and only corresponding values are passed. We also assume |m|≤ℓ and |m'|≤ℓ. Neither of these are checked. The wedge array that this function indexes is ordered as

[
    H(ℓ, m′, m) for ℓ in range(ℓₘₐₓ+1)
    for m′ in range(-min(ℓ, m′ₘₐₓ), min(ℓ, m′ₘₐₓ)+1)
    for m in range(abs(m′), ℓ+1)
]

SphericalFunctions.WignerHsize — Method

WignerHsize(ℓₘₐₓ, m′ₘₐₓ=ℓₘₐₓ)

Total size of array of wedges of width m′ₘₐₓ up to ℓₘₐₓ. If m′ₘₐₓ is not given, it defaults to ℓₘₐₓ.

See also 'WignerHrange` and WignerHindex.

Notes

Here, it is assumed that only data with m≥|m′| are stored, and only corresponding values are passed. We also assume |m|≤ℓ and |m′|≤ℓ. Neither of these are checked. The wedge array that this function indexes is ordered as

[
    H(ℓ, m′, m) for ℓ in 0:ℓₘₐₓ
    for m′ in -min(ℓ, m′ₘₐₓ):min(ℓ, m′ₘₐₓ)
    for m in abs(m′):ℓ
]

SphericalFunctions.Yindex — Function

Yindex(ℓ, m, ℓₘᵢₙ=0)

Compute index into array of mode weights

Parameters

ℓ : int Integer satisfying ℓₘᵢₙ <= ℓ <= ℓₘₐₓ m : int Integer satisfying -ℓ <= m <= ℓ ℓₘᵢₙ : int, optional Integer satisfying 0 <= ℓₘᵢₙ. Defaults to 0.

Returns

i : int Index of a particular element of the mode-weight array as described below

See Also

Ysize : Total size of array of mode weights Yrange : Array of (ℓ, m) indices corresponding to this array

Notes

This assumes that the modes are arranged (with fixed s value) as

[
    Y(s, ℓ, m)
    for ℓ in range(ℓₘᵢₙ, ℓₘₐₓ+1)
    for m in range(-ℓ, ℓ+1)
]

SphericalFunctions.Yrange — Method

Yrange(ℓₘᵢₙ, ℓₘₐₓ)

Create an array of (ℓ, m) indices as in Y array

Parameters

ℓₘᵢₙ : int Integer satisfying 0 <= ℓₘᵢₙ <= ℓₘₐₓ ℓₘₐₓ : int Integer satisfying 0 <= ℓₘᵢₙ <= ℓₘₐₓ

Returns

i : int Total size of array of mode weights arranged as described below

See Also

Ysize : Total size of array of mode weights Yindex : Index of a particular element of the mode weight array

Notes

This assumes that the modes are arranged (with fixed s value) as

[
    Y(s, ℓ, m)
    for ℓ in range(ℓₘᵢₙ, ℓₘₐₓ+1)
    for m in range(-ℓ, ℓ+1)
]

SphericalFunctions.Ysize — Method

Ysize(ℓₘₐₓ)

Compute total size of array of mode weights

See Also

Yrange : Array of (ℓ, m) indices corresponding to this array Yindex : Index of a particular element of the mode weight array

Notes

This assumes that the modes are arranged (with fixed s value) as

[
    Y(s, ℓ, m)
    for ℓ in range(ℓₘᵢₙ, ℓₘₐₓ+1)
    for m in range(-ℓ, ℓ+1)
]

SphericalFunctions._WignerHindex — Method

Helper function for WignerHindex

SphericalFunctions.deduce_limits — Function

deduce_limits(ysize, [ℓmin])

Deduce the value of (ℓmin, ℓmax) that produces $Y$ arrays of the given size.

If ℓmin is not given, it is assumed to be 0. If it is set to nothing, the smallest possible value of ℓmin will be used. However, note that this is not a well-posed problem; multiple combinations of (ℓmin, ℓmax) can give rise to $Y$ arrays of the same size.

See also Ysize

SphericalFunctions.phi_theta — Method

phi_theta(Nϕ, Nθ, [T=Float64])

Construct (phi, theta) grid in order expected by this package.

See also theta_phi for the opposite ordering.

SphericalFunctions.theta_phi — Method

theta_phi(Nθ, Nϕ, [T=Float64])

Construct (theta, phi) grid in spinsfast order.

Note that this order is different from the one assumed by this package; use phi_theta for the opposite ordering.