Utilities

While not usually the star of the show, the following utilities can be quite helpful for actually using the rest of the code.

Complex powers

One common task we find when working with spherical functions is the computation of a range of integer powers of some complex number — so much so that it can be best to pre-compute the powers and cache their values. While a naive approach is generally quite accurate, and reasonably fast, we can do a little better with a specialized routine.

SphericalFunctions.complex_powers! — Method

complex_powers!(zpowers, z)

Compute integer powers of z from z^0 through z^m, recursively, where m is one less than the length of the input zpowers vector.

Note that z is assumed to be normalized, with complex amplitude approximately 1.

See also: complex_powers

SphericalFunctions.complex_powers — Method

complex_powers(z, m)

Compute integer powers of z from z^0 through z^m, recursively.

Note that z is assumed to be normalized, with complex amplitude approximately 1.

This algorithm is mostly due to Stoer and Bulirsch in "Introduction to Numerical Analysis" (page 24) — with a little help from de Moivre's formula, which is essentially exp(iθ)ⁿ = exp(inθ), as well as my own alterations to deal with different behaviors in different quadrants.

There isn't usually a huge advantage to using this specialized function. If you just need a particular power, it will generally be far more efficient and just as accurate to compute either exp(iθ)ⁿ or exp(inθ) explicitly. However, if you need all powers from 0 to m, this function is about 10 or 5 times faster than those options, respectively, for large m. Like those options, this function is numerically stable, in the sense that its errors are usually smaller than m times the error from machine-precision errors in the input argument — or at worst about 50% larger, which occurs as the phase approaches multiples of π/2.

See also: complex_powers!

Sizes of and indexing into $𝔇$, $d$, and $Y$ data

By $Y$ data, we mean anything indexed like $Y_{\ell, m}$ modes; by $D$ data, we mean anything indexed like Wigner's $\mathfrak{D}$ matrices, or special subsets of them, like the $H$ matrices.

SphericalFunctions.WignerDindex — Function

WignerDindex(ℓ, m′, m, m′ₘₐₓ=ℓ)

Compute index into Wigner 𝔇 matrix

See also WignerDrange and WignerDsize.

Notes

This assumes that the Wigner 𝔇 matrix is arranged as

[
    𝔇(ℓ, m′, m)
    for ℓ ∈ ℓₘᵢₙ:ℓₘₐₓ
    for m′ ∈ -min(ℓ, m′ₘₐₓ):min(ℓ, m′ₘₐₓ)
    for m ∈ -ℓ:ℓ
]

SphericalFunctions.WignerDrange — Function

WignerDrange(ℓₘₐₓ, m′ₘₐₓ=ℓₘₐₓ)

Create an array of (ℓ, m', m) indices as in 𝔇 array

See also WignerDsize and WignerDindex.

Notes

This assumes that the Wigner 𝔇 matrix is arranged as

[
    𝔇(ℓ, m′, m)
    for ℓ ∈ ℓₘᵢₙ:ℓₘₐₓ
    for m′ ∈ -min(ℓ, m′ₘₐₓ):min(ℓ, m′ₘₐₓ)
    for m ∈ -ℓ:ℓ
]

SphericalFunctions.WignerDsize — Method

WignerDsize(ℓₘₐₓ, m′ₘₐₓ=ℓₘₐₓ)

Compute total size of Wigner 𝔇 matrix

See also WignerDrange and WignerDindex.

Notes

This assumes that the Wigner 𝔇 matrix is arranged as

[
    𝔇(ℓ, m′, m)
    for ℓ ∈ ℓₘᵢₙ:ℓₘₐₓ
    for m′ ∈ -min(ℓ, m′ₘₐₓ):min(ℓ, m′ₘₐₓ)
    for m ∈ -ℓ:ℓ
]

SphericalFunctions.WignerHindex — Function

WignerHindex(ℓ, m′, m, m′ₘₐₓ)

Index to "wedge" arrays.

See also WignerHsize and WignerHrange.

Notes

Here, it is assumed that only data with m≥|m'| are stored, and only corresponding values are passed. We also assume |m|≤ℓ and |m'|≤ℓ. Neither of these are checked. The wedge array that this function indexes is ordered as

[
    H(ℓ, m′, m) for ℓ ∈ 0:ℓₘₐₓ
    for m′ ∈ -min(ℓ, m′ₘₐₓ):min(ℓ, m′ₘₐₓ)
    for m ∈ abs(m′):ℓ
]

SphericalFunctions.WignerHrange — Function

WignerHrange(ℓₘₐₓ, m′ₘₐₓ=ℓₘₐₓ)

Create an array of (ℓ, m', m) indices as in H array

See also WignerHsize and WignerHindex

Notes

Here, it is assumed that only data with m≥|m'| are stored, and only corresponding values are passed. We also assume |m|≤ℓ and |m'|≤ℓ. Neither of these are checked. The wedge array that this function indexes is ordered as

[
    H(ℓ, m′, m) for ℓ ∈ 0:ℓₘₐₓ
    for m′ ∈ -min(ℓ, m′ₘₐₓ):min(ℓ, m′ₘₐₓ)
    for m ∈ abs(m′):ℓ
]

SphericalFunctions.WignerHsize — Method

WignerHsize(ℓₘₐₓ, m′ₘₐₓ=ℓₘₐₓ)

Total size of array of wedges of width m′ₘₐₓ up to ℓₘₐₓ. If m′ₘₐₓ is not given, it defaults to ℓₘₐₓ.

See also WignerHrange and WignerHindex.

Notes

Here, it is assumed that only data with m≥|m′| are stored, and only corresponding values are passed. We also assume |m|≤ℓ and |m′|≤ℓ. Neither of these are checked. The wedge array that this function indexes is ordered as

[
    H(ℓ, m′, m) for ℓ ∈ 0:ℓₘₐₓ
    for m′ ∈ -min(ℓ, m′ₘₐₓ):min(ℓ, m′ₘₐₓ)
    for m ∈ abs(m′):ℓ
]

SphericalFunctions.Yindex — Function

Yindex(ℓ, m, ℓₘᵢₙ=0)

Compute index into array of mode weights

Parameters

ℓ : int Integer satisfying ℓₘᵢₙ <= ℓ <= ℓₘₐₓ m : int Integer satisfying -ℓ <= m <= ℓ ℓₘᵢₙ : int, optional Integer satisfying 0 <= ℓₘᵢₙ. Defaults to 0.

Returns

i : int Index of a particular element of the mode-weight array as described below

See Also

Ysize : Total size of array of mode weights Yrange : Array of (ℓ, m) indices corresponding to this array

Notes

This assumes that the modes are arranged (with fixed s value) as

[
    Y(s, ℓ, m)
    for ℓ ∈ ℓₘᵢₙ:ℓₘₐₓ
    for m ∈ -ℓ:ℓ
]

SphericalFunctions.Yrange — Method

Yrange(ℓₘᵢₙ, ℓₘₐₓ)

Create an array of (ℓ, m) indices as in Y array

Parameters

ℓₘᵢₙ : int Integer satisfying 0 <= ℓₘᵢₙ <= ℓₘₐₓ ℓₘₐₓ : int Integer satisfying 0 <= ℓₘᵢₙ <= ℓₘₐₓ

Returns

i : int Total size of array of mode weights arranged as described below

See Also

Ysize : Total size of array of mode weights Yindex : Index of a particular element of the mode weight array

Notes

This assumes that the modes are arranged (with fixed s value) as

[
    Y(s, ℓ, m)
    for ℓ ∈ ℓₘᵢₙ:ℓₘₐₓ
    for m ∈ -ℓ:ℓ
]

SphericalFunctions.Ysize — Method

Ysize(ℓₘₐₓ)

Compute total size of array of mode weights

See Also

Yrange : Array of (ℓ, m) indices corresponding to this array Yindex : Index of a particular element of the mode weight array

Notes

This assumes that the modes are arranged (with fixed s value) as

[
    Y(s, ℓ, m)
    for ℓ ∈ ℓₘᵢₙ:ℓₘₐₓ
    for m ∈ -ℓ:ℓ
]

SphericalFunctions._WignerHindex — Method

WignerHindex(ℓ, m′, m, m′ₘₐₓ)

Helper function for WignerHindex, with more constraints.

This function assumes that m ≥ |m′|. The main WignerHindex function uses symmetries of the H array to account for cases that violate this assumption. (But note that both that function and this one assume that |m| ≤ ℓ and |m′| ≤ ℓ.)

SphericalFunctions.deduce_limits — Function

deduce_limits(ysize, [ℓmin])

Deduce the value of (ℓmin, ℓmax) that produces $Y$ arrays of the given size.

If ℓmin is not given, it is assumed to be 0. If it is set to nothing, the smallest possible value of ℓmin will be used. However, note that this is not a well-posed problem; multiple combinations of (ℓmin, ℓmax) can give rise to $Y$ arrays of the same size.

See also Ysize

SphericalFunctions.phi_theta — Method

phi_theta(Nϕ, Nθ, [T=Float64])

Construct (phi, theta) grid in order expected by this package.

See also theta_phi for the opposite ordering.

SphericalFunctions.theta_phi — Method

theta_phi(Nθ, Nϕ, [T=Float64])

Construct (theta, phi) grid in spinsfast order.

Note that this order is different from the one assumed by this package; use phi_theta for the opposite ordering.

Combinatorics

Spherical functions frequently involve binomial coefficients and similar terms, with arguments proportional to $ℓ$, which we aim to allow to be very large — of order 1,000 or more. Unfortunately, due to combinatorical explosions, this is frequently infeasible with naive methods. Here, we collect any specialized methods that help us beat the limits.

SphericalFunctions.sqrtbinomial — Method

sqrtbinomial(n, k, [T])

Evaluate the square-root of the binomial coefficient binomial(n,k) for large coefficients.

Ordinarily, when n and k are standard Int arguments, the built-in binomial function will overflow around n=66, because it results in Ints. We need much larger values. This function, which is based on a related one in SpecialFunctions.jl, returns reasonably accurate results up to n ≈ 1026 when k ≈ n/2 (which is the case of interest in many applications in this package).

Computations are carried out (and returned) in type T, which defaults to Float64.